% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{as.igraph.group_model}
\alias{as.igraph.group_model}
\title{Coerce  a specific group from a \code{group_model} object to an \code{igraph} object.}
\usage{
\method{as.igraph}{group_model}(x, which)
}
\arguments{
\item{x}{The object to convert.}

\item{which}{The number or name of group.}
}
\value{
An \code{igraph} object.
}
\description{
Coerce  a specific group from a \code{group_model} object to an \code{igraph} object.
}
