% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.igraph.matrix}
\alias{as.igraph.matrix}
\title{Coerce a weight matrix to an \code{igraph} object.}
\usage{
\method{as.igraph}{matrix}(x, directed = TRUE, ...)
}
\arguments{
\item{x}{The object to convert.}

\item{...}{Additional arguments. None currently.}

\item{mode}{Either \code{"directed"} or \code{"undirected"}}
}
\value{
An \code{igraph} object.
}
\description{
Coerce a weight matrix to an \code{igraph} object.
}
