% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralities.R
\name{as_centralities}
\alias{as_centralities}
\title{Convert a dataframe to a centralities object}
\usage{
as_centralities(df)
}
\arguments{
\item{df}{A dataframe.}
}
\description{
Convert a dataframe to a centralities object
}
