% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenness.R
\name{betweenness_network}
\alias{betweenness_network}
\alias{betweenness_network.tna}
\title{Build and Visualize a Network with Edge Betweenness}
\usage{
betweenness_network(x, directed = TRUE, invert = TRUE)

\method{betweenness_network}{tna}(x, directed = TRUE, invert = TRUE)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{directed}{A \code{logical} value. If \code{TRUE}, the network is considered
directed.}

\item{invert}{A \code{logical} value indicating whether the weights should be
inverted for distance-based measures. The default is \code{TRUE}.}
}
\value{
A \code{tna} object where the edge weights are edge betweenness values.
}
\description{
This function builds a network from a transition matrix in a \code{tna} object
and computes edge betweenness for the network.
}
\examples{
model <- tna(group_regulation)
betweenness_network(model)

}
\seealso{
Centrality measure functions
\code{\link{centralities}()},
\code{\link{plot.group_tna_centralities}()},
\code{\link{plot.tna_centralities}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.tna_centralities}()}
}
\concept{centralities}
