% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralities.R
\name{calculate_cs}
\alias{calculate_cs}
\title{Calculate Centrality Stability}
\usage{
calculate_cs(corr_mat, threshold, certainty, drop_prop)
}
\arguments{
\item{corr_mat}{A \code{matrix} of correlation values}

\item{threshold}{A \code{numeric} value specifying the correlation threshold for
calculating the CS-coefficient. THe default is 0.7.}

\item{certainty}{A \code{numeric} value specifying the desired level of certainty
for the CS-coefficient. Default is 0.95.}

\item{drop_prop}{A \code{numeric} vector specifying the proportions of
cases to drop in each sampling iteration. Default is a sequence from 0.1 to
0.9 in increments of 0.1.}
}
\description{
Calculate Centrality Stability
}
