% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{calculate_network_metrics}
\alias{calculate_network_metrics}
\title{Calculate Network Metrics for a Cluster}
\usage{
calculate_network_metrics(x, cluster = 1)
}
\arguments{
\item{x}{A \code{tna} object}

\item{cluster}{An integer specifying which cluster to analyze. Defaults to \code{1}.}
}
\value{
A named list containing the following network metrics:
\itemize{
\item \code{Node_count}: The total number of nodes.
\item \code{Edge_count}: The total number of edges.
\item \code{Network_Density}: The density of the network.
\item \code{Mean_distance}: The mean shortest path length.
\item \code{Mean_out_strength}: The mean out-strength of nodes.
\item \code{SD_out_strength}: The standard deviation of out-strength.
\item \code{Mean_in_strength}: The mean in-strength of nodes.
\item \code{SD_in_strength}: The standard deviation of in-strength.
\item \code{Mean_out_degree}: The mean out-degree of nodes.
\item \code{SD_out_degree}: The standard deviation of out-degree.
\item \code{Centralization_out_degree}: The centralization of out-degree.
\item \code{Centralization_in_degree}: The centralization of in-degree.
\item \code{Reciprocity}: The reciprocity of the network.
}
}
\description{
This function calculates a variety of network metrics for a specified cluster
in a transition network stored in a \code{tna} object. It computes key metrics such
as node and edge counts, network density, mean distance, strength measures,
degree centrality, and reciprocity. A histogram of edge weights is also plotted.
}
\details{
The function extracts the \code{igraph} network for the specified cluster and computes the following network metrics:
\itemize{
\item Node count: Total number of nodes in the network.
\item Edge count: Total number of edges in the network.
\item Network density: Proportion of possible edges that are present in the network.
\item Mean distance: The average shortest path length between nodes.
\item Mean and standard deviation of out-strength and in-strength: Measures of the total weight of outgoing and incoming edges for each node.
\item Mean and standard deviation of out-degree: The number of outgoing edges from each node.
\item Centralization of out-degree and in-degree: Measures of how centralized the network is based on the degrees of nodes.
\item Reciprocity: The proportion of edges that are reciprocated (i.e., mutual edges between nodes).
}

A summary of the metrics is printed to the console, and a histogram of edge weights (probabilities) is plotted.
}
\examples{
\dontrun{
# Calculate network metrics for the first cluster
network_metrics <- calculate_network_metrics(tna_model, cluster = 1)
}
}
\seealso{
\code{igraph}
}
