% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques.R
\name{cliques}
\alias{cliques}
\alias{cliques.tna}
\title{Identify Cliques in a Transition Network}
\usage{
cliques(x, ...)

\method{cliques}{tna}(x, size = 3, threshold = 0, sum_weights = FALSE, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{...}{Ignored.}

\item{size}{An \code{integer} specifying the size of the cliques to identify.
Defaults to \code{3} (triads).}

\item{threshold}{A \code{numeric} value that sets the minimum edge weight
for an edge to be considered in the clique. Edges below this value
are ignored. Defaults to \code{0}.}

\item{sum_weights}{A \code{logical} value specifying whether the sum of the
weights should be above the \code{threshold} instead of individual weights of the
directed edges. Defaults to \code{FALSE}.}
}
\value{
A \code{tna_cliques} object which is a \code{list} of two elements:
\itemize{
\item \code{weights} is a \code{matrix} of the edge weights in the clique.
\item \code{inits} is a \code{numeric} vector of initial weights for the clique.
}
}
\description{
This function identifies cliques of a specified size in a transition network.
It searches for cliques—complete subgraphs where every pair of nodes is
connected—of size \code{n} in the transition matrix for the specified cluster
in the \code{tna} object.
}
\examples{
model <- tna(engagement)

# Find  2-cliques (dyads)
cliq <- cliques(model, size = 2)

}
