% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R, R/print.R
\name{cluster_sequences}
\alias{cluster_sequences}
\alias{print.tna_clustering}
\title{Cluster Sequences via Dissimilarity Matrix based on String Distances}
\usage{
cluster_sequences(
  data,
  k,
  dissimilarity = "osa",
  method = "pam",
  na_syms = c("*", "\%"),
  ...
)

\method{print}{tna_clustering}(x, ...)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{matrix} where the rows are sequences and
the columns are time points.}

\item{k}{An \code{integer} giving the number of clusters.}

\item{dissimilarity}{A \code{character} string specifying the
dissimilarity measure. The available options are: \code{"osa"}, \code{"lv"}, \code{"dl"},
\code{"hamming"}, \code{"qgram"}, \code{"cosine"}, \code{"jaccard"}, and \code{"jw"}. See
\link[stringdist:stringdist-metrics]{stringdist::stringdist-metrics} for more information on these measures.}

\item{method}{A \code{character} string specifying clustering method.
The available methods are \code{"pam"}, \code{"ward.D"}, \code{"ward.D2"},
\code{"complete"},\code{"average"}, \code{"single"}, \code{"mcquitty"}, \code{"median"}, and
\code{"centroid"}. See \code{\link[cluster:pam]{cluster::pam()}} and \code{\link[stats:hclust]{stats::hclust()}} for more
information on these methods.}

\item{...}{Additional arguments passed to \code{\link[stringdist:stringdist]{stringdist::stringdist()}}.}

\item{x}{A \code{tna_clustering} object.}
}
\value{
A \code{tna_clustering} object which is a  \code{list} containing:
\itemize{
\item \code{data}: The original data.
\item \code{k}: The number of clusters.
\item \code{assignments}: An \code{integer} vector of cluster assignments.
\item \code{silhouette}: Silhouette score measuring clustering quality.
\item \code{sizes}: An \code{integer} vector of cluster sizes.
\item \code{method}: The clustering method used.
\item \code{distance}: The distance matrix.
}
}
\description{
Performs clustering on sequence data using specified dissimilarity measures
and clustering methods. The sequences are first converted to strings
and compared using the \code{stringdist} package.
}
\examples{
data <- data.frame(
  T1 = c("A", "B", "A", "C", "A", "B"),
  T2 = c("B", "A", "B", "A", "C", "A"),
  T3 = c("C", "C", "A", "B", "B", "C")
)

# PAM clustering with optimal string alignment (default)
result <- cluster_sequences(data, k = 2)
print(result)

}
