% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.tna_mmm}
\alias{coef.tna_mmm}
\title{Extract Regression Coefficients of a Mixture Markov Model}
\usage{
\method{coef}{tna_mmm}(object, ...)
}
\arguments{
\item{object}{A \code{tna_mmm} object.}

\item{...}{Not used.}
}
\value{
A \code{matrix} containing the regression coefficient estimates.
}
\description{
Extract Regression Coefficients of a Mixture Markov Model
}
\examples{
coef(engagement_tna_mmm)

}
