% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.group_tna}
\alias{compare.group_tna}
\title{Compare TNA Clusters with Comprehensive Metrics}
\usage{
\method{compare}{group_tna}(x, i = 1L, j = 2L, scaling = "none", ...)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{i}{An \code{integer} index or the name of the principal cluster as a
\code{character} string.}

\item{j}{An \code{integer} index or the name of the secondary cluster as a
\code{character} string.}

\item{scaling}{See \code{\link[=compare.tna]{compare.tna()}}.}

\item{...}{Additional arguments passed to \code{\link[=compare.tna]{compare.tna()}}.}
}
\value{
A \code{tna_comparison} object. See \code{\link[=compare.tna]{compare.tna()}} for details.
}
\description{
Compare TNA Clusters with Comprehensive Metrics
}
\examples{
model <- group_model(engagement_mmm)
compare(model, i = 1, j = 2)

}
