% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare Two Matrices or Models with Comprehensive Metrics and Visualizations}
\usage{
compare(x, y, ...)
}
\arguments{
\item{x}{A \code{tna} object or a \code{matrix} of weights.}

\item{y}{A \code{tna} object or a \code{matrix} of weights.}

\item{...}{Ignored.}
}
\value{
A \code{tna_compare} object, which is a \code{list} containing the
following elements:
\itemize{
\item \code{matrices}: A \code{list} containing the original matrices of the input \code{tna}
objects or the inputs themselves in case of matrices.
\item \code{difference_matrix}: A \code{matrix} of differences \code{x - y}.
\item \code{edge_metrics}: A \code{data.frame} of edge-level metrics about the differences.
\item \code{summary_metrics}: A \code{data.frame} of summary metrics of the differences
across all edges.
\item \code{centrality_differences}: A \code{data.frame} of differences in centrality
measures computes from \code{x} and \code{y}.
\item \code{centrality_correlations}: A \code{numeric} vector of correlations of the
centrality measures between \code{x} and \code{y}.
}
}
\description{
Compare Two Matrices or Models with Comprehensive Metrics and Visualizations
}
\examples{
# Create two example matrices
matrix1 <- matrix(runif(100, 0, 1), nrow = 10, ncol = 10)
matrix2 <- matrix(runif(100, 0, 1), nrow = 10, ncol = 10)
rownames(matrix1) <- rownames(matrix2) <- paste0("Gene", 1:10)
colnames(matrix1) <- colnames(matrix2) <- paste0("Sample", 1:10)
result <- compare(matrix1, matrix2)

}
