% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{compare_tna_networks}
\alias{compare_tna_networks}
\title{Compare Two Networks from Sequence Data Using Permutation Tests}
\usage{
compare_tna_networks(
  x,
  y,
  cluster1 = 1,
  cluster2 = 1,
  iter = 1000,
  paired = FALSE,
  alpha = 0.05,
  progressbar = TRUE
)
}
\arguments{
\item{x}{A \code{tna} object containing sequence data for the first \code{tna} model.}

\item{y}{A \code{tna} object containing sequence data for the second \code{tna} model.}

\item{cluster1}{An integer indicating the cluster to be used from the first \code{tna} model. Default is 1.}

\item{cluster2}{An integer indicating the cluster to be used from the second \code{tna} model. Default is 1.}

\item{iter}{The number of permutations to perform. Default is 1000.}

\item{paired}{Logical. If \code{TRUE}, perform paired permutation tests; if \code{FALSE}, perform unpaired tests. Default is \code{FALSE}.}

\item{alpha}{The significance level for the permutation tests. Default is 0.05.}

\item{progressbar}{Logical. If \code{TRUE}, display a progress bar during the permutation process. Default is \code{TRUE}.}
}
\value{
A list containing:
\itemize{
\item \code{edge_statistics} A data frame with edge names, original differences, p-values, effect sizes, and confidence intervals.
\item \code{original_matrix1} The transition probability matrix from the first \code{tna} model
\item \code{original_matrix2} The transition probability matrix from the second \code{tna} model.
\item \code{difference_matrix} The matrix of differences in transition probabilities.}
\item \code{significant_matrix} A matrix showing the significant edges based on the permutation tests.
\item \code{significant_edges} A list of significant edges (p < alpha).
\item \code{statistics_summary} A list of summary statistics including mean absolute differences and effect sizes.
}
}
\description{
This function compares two networks built from sequence data using permutation tests.
The function builds Markov models for two sequence objects , computes the transition
probabilities, and compares them by performing permutation tests. It returns the
differences in transition probabilities, effect sizes, p-values, and confidence intervals.
}
\examples{
\dontrun{
# Assuming `network1` and `network2` are tna objects containing sequence data
result <- compare_tna_networks(network1, network2, cluster1 = 1, cluster2 = 1, it = 1000)
}

}
\seealso{
Evaluation and validation functions
\code{\link{bootstrap}()},
\code{\link{prune}()}
}
\concept{evaluation}
