% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R, R/pruning.R
\name{deprune.group_tna}
\alias{deprune.group_tna}
\alias{deprune}
\alias{deprune.tna}
\alias{reprune.tna}
\title{Restore a Pruned Transition Network Analysis Model}
\usage{
\method{deprune}{group_tna}(x, ...)

deprune(x, ...)

\method{deprune}{tna}(x, ...)

\method{reprune}{tna}(x, ...)
}
\arguments{
\item{x}{A \code{tna} or \code{group_tna} object.}

\item{...}{Ignored.}
}
\description{
Restore a Pruned Transition Network Analysis Model
}
\examples{
model <- tna(engagement)
pruned_model <- prune(model, method = "threshold", threshold = 0.1)
depruned_model <- deprune(pruned_model) # restore original model
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{prune.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
