% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{dissimilarity_matrix}
\alias{dissimilarity_matrix}
\title{Dissimilarity matrix}
\usage{
dissimilarity_matrix(strings, dissimilarity, lambda, ...)
}
\arguments{
\item{strings}{Output of \code{seq2str}.}

\item{lambda}{A \code{numeric} value for the decay rate.}
}
\description{
Dissimilarity matrix
}
