% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{engagement}
\alias{engagement}
\title{Example data on student engagement}
\format{
An object of class \code{stslist} (inherits from \code{data.frame}) with 200 rows and 20 columns.
}
\usage{
engagement
}
\description{
TODO explain more
}
\keyword{datasets}
