% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{engagement_tna_mmm}
\alias{engagement_tna_mmm}
\title{Example Mixed Markov Model Fitted to the \code{engagement} Data via TNA}
\format{
A \code{tna_mmm} object.
}
\source{
The data was generated via \code{mixed_markov_model.R} in
\url{https://github.com/sonsoleslp/tna/tree/main/data-raw/}
}
\usage{
engagement_tna_mmm
}
\description{
Example Mixed Markov Model Fitted to the \code{engagement} Data via TNA
}
\seealso{
Datasets
\code{\link{engagement}},
\code{\link{engagement_mmm}},
\code{\link{group_regulation}},
\code{\link{group_regulation_long}}
}
\concept{datasets}
\keyword{datasets}
