% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralities.R
\name{estimate_cs}
\alias{estimate_cs}
\alias{estimate_centrality_stability}
\title{Estimate Centrality Stability}
\usage{
estimate_cs(
  x,
  measures = c("InStrength", "OutStrength", "Betweenness"),
  normalize = FALSE,
  iter = 1000,
  method = "pearson",
  drop_prop = seq(0.1, 0.9, by = 0.1),
  threshold = 0.7,
  certainty = 0.95,
  detailed = FALSE
)

estimate_centrality_stability(
  x,
  measures = c("InStrength", "OutStrength", "Betweenness"),
  normalize = FALSE,
  iter = 1000,
  method = "pearson",
  drop_prop = seq(0.1, 0.9, by = 0.1),
  threshold = 0.7,
  certainty = 0.95,
  detailed = FALSE
)
}
\arguments{
\item{x}{A \code{tna} object representing the temporal network analysis data.
The object should be created from a sequence datad object.}

\item{measures}{A \code{character} vector of centrality measures to estimate.
The default measures are \code{"InStrength"}, \code{"OutStrength"},
and \code{"Betweenness"}.}

\item{normalize}{A \code{logical} value indicating whether to normalize
the centrality measures. The default is \code{FALSE}.}

\item{iter}{An \code{integer} specifying the number of resamples to draw.
The default is 1000.}

\item{method}{A \code{character} string indicating the correlation coefficient
type. The default is \code{"pearson"}. See \code{\link[stats:cor]{stats::cor()}} for details.}

\item{drop_prop}{A \code{numeric} vector specifying the proportions of
cases to drop in each sampling iteration. Default is a sequence from 0.1 to
0.9 in increments of 0.1.}

\item{threshold}{A \code{numeric} value specifying the correlation threshold for
calculating the CS-coefficient. THe default is 0.7.}

\item{certainty}{A \code{numeric} value specifying the desired level of certainty
for the CS-coefficient. Default is 0.95.}

\item{detailed}{A \code{logical} value specifying whether to return detailed
sampling results. If \code{TRUE}, detailed results are included in the output.
THe default is \code{FALSE}.}
}
\value{
A \code{tna_stability} object which is a \code{list} with an element for each
\code{measure} with the following elements:
\itemize{
\item \code{cs_coefficient}: The centrality stability (CS) coefficient
of the measure.
\item \code{correlations}: A \code{matrix} of correlations between the original
centrality and the resampled centralities for each drop proportion.
\item \code{detailed_results}: A detailed data frame of the sampled correlations,
returned only if \code{return_detailed = TRUE}.
}
}
\description{
This function estimates the stability of centrality measures in a network
using subset sampling without replacement. It allows for dropping varying
proportions of cases and calculates correlations between the original
centralities and sampled subsets.
}
\details{
The function works by repeatedly resampling the data, dropping varying
proportions of cases, and calculating centrality measures on the subsets.
The correlation between the original centralities and the resampled
centralities is calculated for each drop proportion. The stability of each
centrality measure is then summarized using a centrality stability (CS)
coefficient, which represents the proportion of dropped cases at which
the correlations drop below a given threshold (default 0.7).

The results can be visualized with an optional plot showing the stability of
the centrality measures across different drop proportions, along with
confidence intervals. The CS-coefficients are displayed in the plot's
subtitle.
}
\examples{
model <- tna(engagement)
estimate_cs(model, measures = c("InStrength", "OutStrength"), iter = 10)

}
