% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{event2sequence}
\alias{event2sequence}
\title{Parse Event Data Into Sequence Data}
\usage{
event2sequence(
  data,
  actor_col,
  time_col,
  action_col,
  time_threshold = 900,
  custom_format = NULL,
  is_unix_time = FALSE,
  unix_time_unit = "seconds",
  verbose = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} with three columns for the actor, time, and
action.}

\item{actor_col}{A \code{character} string naming the actor column.}

\item{time_col}{A \code{character} string naming the actor column.}

\item{action_col}{A \code{character} string naming the actor column.}

\item{time_threshold}{TODO}

\item{custom_format}{A \code{character} string giving the format used to
parse the time column.}

\item{is_unix_time}{A \code{logical} value indicating whether the time column
is in Unix time. The default is \code{FALSE}.}

\item{unix_time_unit}{A \code{character} string giving the Unix time unit. The
default is \code{"seconds"},}

\item{verbose}{A \code{logical} value indicating whether to print informative
messages during the parsing process. The default is \code{TRUE}.}
}
\value{
TODO.
}
\description{
Parse Event Data Into Sequence Data
}
\examples{
# TODO

}
