% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{find_communities}
\alias{find_communities}
\title{Internal Community Detection Function}
\usage{
find_communities(g, Gamma = 1)
}
\arguments{
\item{g}{An \code{igraph} object representing the transition network.}

\item{Gamma}{A numeric parameter that influences the Spin Glass algorithm. Defaults to \code{1}.}
}
\value{
A list containing:
\itemize{
\item \code{community_counts}: A list with the number of communities found by each algorithm.
\item \code{community_assignments}: A dataframe where each row corresponds to a
node and each column to a community detection algorithm, with color-coded
community assignments.
}
}
\description{
This function detects communities in a given graph using various algorithms
provided by the \code{igraph} package.
}
\details{
This function applies multiple community detection algorithms (e.g., Walktrap,
Fast Greedy, Label Propagation, Infomap, Edge Betweenness, Leading Eigenvector,
and Spin Glass) to the input graph \code{g}. It then constructs a dataframe of
community assignments and a list of the number of communities detected by each algorithm.
}
\author{
Mohammed Saqr (\email{mohammed.saqr@uef.fi})
}
