% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find_dyads}
\alias{find_dyads}
\title{Find Dyads in a Transition Network}
\usage{
find_dyads(
  x,
  cluster = 1,
  weight_threshold = 0,
  directed = FALSE,
  sum_weights = FALSE,
  minimum = 0.00001,
  mar = c(10, 10, 10, 10),
  ...
)
}
\arguments{
\item{x}{A \code{tna} object that contains transition matrices and associated node metadata.}

\item{cluster}{An integer specifying which cluster to analyze. Defaults to \code{1}.}

\item{weight_threshold}{A numeric value that sets the minimum weight for an
edge to be considered. Edges below this threshold are ignored. Defaults to \code{0}.}

\item{directed}{A logical value indicating whether the network is directed.
Defaults to \code{FALSE}.}

\item{sum_weights}{A logical value specifying whether to sum weights for
directed edges when calculating triangle weights. Defaults to \code{FALSE}.}

\item{minimum}{A numeric value for the minimum edge weight to be displayed
in the visualization. Defaults to \code{0.00001}.}

\item{mar}{A numeric vector specifying the margins for the visualization plot.}

\item{...}{Additional parameters passed to the \code{qgraph} function for plotting.}
}
\value{
A list with two elements:
\itemize{
\item \code{count}: The total number of dyads found.
\item \code{dyads}: A list of detected dyads, where each element contains:
\itemize{
\item \code{nodes}: The labels of the nodes forming the dyad
\item \code{weights}: The edge weights between the nodes.
\item \code{matrix}: The submatrix representing the connections between the
nodes in the dyad
\item \code{pie_values}: The initial state probabilities (\code{Pie} values) for
the nodes in the dyad.
}
}
}
\description{
This function identifies and visualizes dyads (sets of two nodes forming a pair)
in a transition network from a \code{tna} object. It can handle both directed
and undirected networks, and it allows for customizable visualization.
}
\details{
This function loops through all possible pairs of nodes in the network
and checks whether the edges between them form a complete dyad
The \code{weight_threshold} parameter allows filtering out weak edges, and the
\code{sum_weights} option sums directed edge weights.

For each detected dyad, the function stores the nodes, their weights,
the submatrix corresponding to the dyad, and the initial state
probabilities (\code{Pie} values) for the nodes. It then visualizes each dyad
using the \code{qgraph} package, allowing customization of the plot via additional
arguments.
}
\examples{
\dontrun{
# Find and visualize dyads in the first cluster
dyads <- find_dyads(tna_model, cluster = 1)
}

}
\seealso{
\code{qgraph}
}
\author{
Mohammed Saqr (\email{mohammed.saqr@uef.fi})
}
