% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find_dyads_clusters}
\alias{find_dyads_clusters}
\title{Find Dyads Across All Clusters}
\usage{
find_dyads_clusters(x, ...)
}
\arguments{
\item{x}{A \code{tna} object containing transition matrices for multiple clusters.}

\item{...}{Additional arguments passed to the \code{find_dyads} function.}
}
\value{
A list where each element contains the dyads identified for a specific
cluster. Each cluster's result is the output of the \code{find_dyads} function.
}
\description{
This function identifies dyads (sets of two nodes) in all transition matrices for each cluster within a \code{tna} object. It loops through each cluster and applies the \code{find_dyads} function to detect dyads.
}
\details{
The function iterates over the transition matrices stored in the \code{tna} object
for each cluster. For each cluster, it prints a message indicating the cluster
being processed and then applies the \code{find_dyads} function to detect dyads
in that cluster.
}
\examples{
\dontrun{
# Find dyads across all clusters in a `tna` object
dyads_result <- find_dyads_clusters(tna_model)
}
}
\seealso{
\code{find_dyads}
}
\author{
Sonsoles López-Pernas (\email{sonsoles.lopez@uef.fi})
}
