% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find_triads_clusters}
\alias{find_triads_clusters}
\title{Find Triads Across All Clusters}
\usage{
find_triads_clusters(x, ...)
}
\arguments{
\item{x}{A \code{tna} object containing transition matrices for multiple clusters.}

\item{...}{Additional arguments passed to the \code{find_triads} function.}
}
\value{
A list where each element contains the triads identified for a specific
cluster. Each cluster's result is the output of the \code{find_triads} function.
}
\description{
This function identifies triangles or triads (sets of thre nodes) in all
transition matrices for each cluster within a \code{tna} object. It loops through
each cluster and applies the \code{find_triads} function to detect triads
}
\details{
The function iterates over the transition matrices stored in the \code{tna} object
for each cluster. For each cluster, it prints a message indicating the cluster
being processed and then applies the \code{find_triads} function to detect triads in
that cluster.
}
\examples{
\dontrun{
# Find triads across all clusters in a `tna` object
triads_result <- find_triads_clusters(tna_model)
}
}
\seealso{
\code{find_triads}

Pattern-finding functions
\code{\link{find_communities}()},
\code{\link{find_dyads}()},
\code{\link{find_dyads_clusters}()},
\code{\link{find_triads}()},
\code{\link{plot_communities}()}
}
\author{
Sonsoles López-Pernas (\email{sonsoles.lopez@uef.fi})
}
\concept{patterns}
