% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_model}
\alias{get_model}
\title{Get the Model from TNA Results}
\usage{
get_model(results)
}
\arguments{
\item{results}{An object containing the results of a network analysis,
including the model and other relevant information.}
}
\value{
The  \code{tna}  model object extracted from the results.
}
\description{
This function extracts and returns the \code{tna} model object from a  results list
The \code{tna} model is typically stored within the results of an analysis
conducted with the \code{tna} package.
}
\examples{
\dontrun{
# Assuming `results` is a tna object
model <- get_model(results)
}
}
