% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{group_regulation}
\alias{group_regulation}
\title{Example data on group regulation}
\format{
An object of class \code{data.frame} with 2000 rows and 26 columns.
}
\usage{
group_regulation
}
\description{
Students' regulation during collaborative learning. Students' interactions
were coded as:  "adapt", "cohesion", "consensus", "coregulate", "discuss",
"emotion", "monitor", "plan", "synthesis"
The data was generated synthetically.
}
\keyword{datasets}
