% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{group_regulation_long}
\alias{group_regulation_long}
\title{Example Long Data on Group Regulation}
\format{
A \code{data.frame} object.
}
\source{
The data was generated synthetically from \code{group_regulation}
}
\usage{
group_regulation_long
}
\description{
Students' regulation during collaborative learning. This is the same dataset
as \code{group_regulation} but in long format. In addition to students'
actions (\code{Action}), it contains the student identifier (\code{Actor}),
timestamp (\code{Time}), \code{Course} name, and collaboration \code{Group}. It
also includes a column (\code{Achiever}) indicating whether the student
is a high or low achiever.
}
\seealso{
Datasets
\code{\link{engagement}},
\code{\link{engagement_mmm}},
\code{\link{group_regulation}}
}
\concept{datasets}
\keyword{datasets}
