% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.group_tna}
\alias{hist.group_tna}
\title{Plot a Histogram of Edge Weights for a \code{group_tna} Object.}
\usage{
\method{hist}{group_tna}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{...}{Additional arguments passed to \code{\link[graphics:hist]{graphics::hist()}}.}
}
\value{
A \code{list} (invisibly) of \code{histogram} objects of the edge weights of
each cluster.
}
\description{
Plot a Histogram of Edge Weights for a \code{group_tna} Object.
}
\examples{
model <- group_model(engagement_mmm)
hist(model)

}
\seealso{
Cluster-related functions
\code{\link{bootstrap}()},
\code{\link{centralities}()},
\code{\link{cliques}()},
\code{\link{communities}()},
\code{\link{deprune}()},
\code{\link{estimate_cs}()},
\code{\link{group_model}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.group_tna_centralities}()},
\code{\link{plot.group_tna_cliques}()},
\code{\link{plot.group_tna_communities}()},
\code{\link{plot.group_tna_stability}()},
\code{\link{plot_compare.group_tna}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_cliques}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.group_tna_stability}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{prune}()},
\code{\link{pruning_details}()},
\code{\link{rename_groups}()},
\code{\link{reprune}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.group_tna_bootstrap}()}
}
\concept{clusters}
