% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{import_ts}
\alias{import_ts}
\title{Import and Convert Time-Series Data into Wide Format Sequence Data}
\usage{
import_ts(
  data,
  id_col,
  value_col,
  n_states,
  method = "kmeans",
  unused_fn = dplyr::first,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing time-series data in long format.}

\item{id_col}{An optional \code{character} string naming the column that contains
the unique IDs.}

\item{value_col}{A \code{character} string naming the column that contains the
data values.}

\item{n_states}{An \code{integer} specifying the number of states.}

\item{method}{A \code{character} string defining the discretization method to use.
The available options are:
\itemize{
\item \code{width}: for equal width binning.
\item \code{quantile}: for quantile-based binning.
\item \code{kde}: for binning based on kernel density estimation.
\item \code{gaussian}: for a Gaussian mixture model.
\item \code{kmeans}: for K-means clustering (the default).
}}

\item{unused_fn}{How to handle extra columns when pivoting to wide format.
See \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}. The default is to keep all columns and to
use the first value.}

\item{...}{Additional arguments passed to the discretization method
(\code{\link[stats:kmeans]{stats::kmeans()}} for \code{kmeans}, \code{\link[stats:density]{stats::density()}} and
\code{\link[pracma:findpeaks]{pracma::findpeaks()}} for \code{kde}, and \code{\link[mclust:Mclust]{mclust::Mclust()}} for \code{gaussian}).}
}
\value{
A \code{data.frame} in wide format containing the sequence data.
}
\description{
Imports time-series data as sequence data by discretizing the input data.
Various methods for discretization are available including gaussian mixtures,
K-means clustering and kernel density based binning.
}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{import_data}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{prepare_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{print.tna_data}()},
\code{\link{simulate.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
