% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{mmm_stats}
\alias{mmm_stats}
\title{Retrieve statistics from a mixture Markov model (MMM)}
\usage{
mmm_stats(x, use_t_dist = TRUE, level = 0.05)
}
\arguments{
\item{x}{A \code{mhmm} object.}

\item{use_t_dist}{A \code{logical} value. If \code{TRUE} (the default), the
t-distribution is used to compute confidence intervals.}

\item{level}{A \code{numeric} value representing the significance level for
hypothesis testing and confidence intervals. Defaults to \code{0.05}.}
}
\value{
A \code{data.frame} object.
}
\description{
Retrieve statistics from a mixture Markov model (MMM)
}
\examples{
mmm_stats(engagement_mmm)

}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities}()},
\code{\link{deprune}()},
\code{\link{estimate_cs}()},
\code{\link{group_model}()},
\code{\link{prune}()},
\code{\link{pruning_details}()},
\code{\link{rename_groups}()},
\code{\link{reprune}()}
}
\concept{clusters}
