% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation_test}
\alias{permutation_test}
\title{Compare Two Networks from Sequence Data Using Permutation Tests}
\usage{
permutation_test(
  x,
  y,
  cluster1 = 1,
  cluster2 = 1,
  iter = 1000,
  paired = FALSE,
  level = 0.05
)
}
\arguments{
\item{x}{A \code{tna} object containing sequence data for the first \code{tna} model.}

\item{y}{A \code{tna} object containing sequence data for the second \code{tna} model.}

\item{cluster1}{An \code{integer} indicating the cluster to be used for
the first \code{tna} model. Default is 1.}

\item{cluster2}{An \code{integer} indicating the cluster to be used for
the second \code{tna} model. Default is 1.}

\item{iter}{The number of permutations to perform. Default is 1000.}

\item{paired}{Logical. If \code{TRUE}, perform paired permutation tests;
if \code{FALSE}, perform unpaired tests. Default is \code{FALSE}.}

\item{level}{The significance level for the permutation tests.
The default is 0.05.}
}
\value{
A \code{list} containing:
\itemize{
\item \code{edge_statistics} A data frame with edge names, original differences,
and p-values.
\item \code{original_weights1} The transition probability matrix from
the first \code{tna} model
\item \code{original_weights2} The transition probability matrix from
the second \code{tna} model.
\item \code{true_diffs} The matrix of differences in weights.
\item \code{sig_diffs} A matrix showing the significant differences
based on the permutation test.
}
}
\description{
This function compares two networks built from sequence data using permutation tests.
The function builds Markov models for two sequence objects , computes the transition
probabilities, and compares them by performing permutation tests. It returns the
differences in transition probabilities, effect sizes, p-values, and confidence intervals.
}
\examples{
\dontrun{
result <- permutation_test(network1, network2, cluster1 = 1, cluster2 = 1, it = 1000)
}

}
\seealso{
Evaluation and validation functions
\code{\link{bootstrap}()},
\code{\link{prune}()}
}
\concept{evaluation}
