% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralities.R
\name{plot.centralities}
\alias{plot.centralities}
\title{Plot Centralities for a Transition Matrix}
\usage{
plot.centralities(mat)
}
\arguments{
\item{mat}{A square matrix representing transition probabilities or adjacency.}
}
\value{
A ggplot object displaying the lollipop charts for each centrality measure.
}
\description{
This function plots various centrality measures for a given transition matrix using a lollipop chart.
}
\details{
This function calculates several centrality measures using the \code{\link{calculate.centralities}} function and then plots these measures using the `lollipop_chart` from `ggcharts`.
The centrality measures include OutStrength, InStrength, ClosenessIn, ClosenessOut, Closeness, Betweenness, Diffusion, and Clustering.
The resulting plot includes facets for each centrality measure, showing the values for each interaction. The returned plot is a `ggplot2` object, so it can be easily modified and styled.
}
\examples{
\dontrun{
  library(TraMineR)
  data(biofam3c)

  # Preparing the sequence data
  seq_data <- seqdef(biofam3c$biofam)

  # Building a transition matrix from the sequence data
  tna_model <- build.tna(seq_data)
  transition_matrix <- tna_model$Matrix

  # Plotting the centralities
  plot.centralities(transition_matrix)
}

}
