% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.group_tna_bootstrap}
\alias{plot.group_tna_bootstrap}
\title{Plot a Bootstrapped Grouped Transition Network Analysis Model}
\usage{
\method{plot}{group_tna_bootstrap}(x, title = names(x), ...)
}
\arguments{
\item{x}{A \code{group_tna_bootstrap} object.}

\item{title}{A \code{character} vector of titles to use for each plot.}

\item{...}{Additional arguments passed to \code{\link[=plot.tna]{plot.tna()}}.}
}
\description{
Plot a Bootstrapped Grouped Transition Network Analysis Model
}
\examples{
model <- group_model(engagement_mmm)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 50)
plot(boot)

}
\seealso{
Validation functions
\code{\link{bootstrap}()},
\code{\link{deprune}()},
\code{\link{estimate_cs}()},
\code{\link{permutation_test}()},
\code{\link{permutation_test.group_tna}()},
\code{\link{plot.group_tna_permutation}()},
\code{\link{plot.group_tna_stability}()},
\code{\link{plot.tna_bootstrap}()},
\code{\link{plot.tna_permutation}()},
\code{\link{plot.tna_stability}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_permutation}()},
\code{\link{print.group_tna_stability}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{print.summary.tna_bootstrap}()},
\code{\link{print.tna_bootstrap}()},
\code{\link{print.tna_permutation}()},
\code{\link{print.tna_stability}()},
\code{\link{prune}()},
\code{\link{pruning_details}()},
\code{\link{reprune}()},
\code{\link{summary.group_tna_bootstrap}()},
\code{\link{summary.tna_bootstrap}()}
}
\concept{validation}
