% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.group_tna_communities}
\alias{plot.group_tna_communities}
\title{Plot Detected Communities}
\usage{
\method{plot}{group_tna_communities}(x, title = names(x), colors, ...)
}
\arguments{
\item{x}{A \code{group_tna_communities} object.}

\item{title}{A \code{character} vector of titles to use for each plot.}

\item{colors}{A \code{character} vector of colors to use.}

\item{...}{Arguments passed to \code{\link[=plot.tna_communities]{plot.tna_communities()}}.}
}
\value{
A \code{list} (invisibly) of \code{qgraph} objects in which the nodes are
colored by community for each cluster.
}
\description{
Plot Detected Communities
}
\examples{
model <- group_model(engagement_mmm)
comm <- communities(model)
plot(comm)

}
\seealso{
Community detection functions
\code{\link{communities}()},
\code{\link{plot.tna_communities}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.tna_communities}()}
}
\concept{communities}
