% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{plot.group_tna}
\alias{plot.group_tna}
\title{Plot a grouped Transition Network Analysis Model}
\usage{
\method{plot}{group_tna}(x, title, ...)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{title}{A title for each plot. It can be a single string (the same one
will be used for all plots) or a list (one per group)}

\item{...}{Same as \code{\link[=plot.tna]{plot.tna()}}}
}
\description{
Plot a grouped Transition Network Analysis Model
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{prune.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
