% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna_centralities}
\alias{plot.tna_centralities}
\title{#' Plot Transition Networks for All Clusters
#'
#' This function plots the transition networks for each cluster in a \code{tna} object.
#' It iterates through the transition matrices for each cluster and generates
#' corresponding plots using the \code{plot.tna} function.
#'
#' @param x A \code{tna} object containing transition matrices for different clusters.
#' @param ... Additional arguments to be passed to the \code{plot.tna} function.
#'
#' @return A series of plots showing the transition networks for each cluster.
#' @family core
#' @examples
#' \dontrun{
#' # Assuming `tna_model` is a tna object containing transition matrices
#' plot_clusters(tna_model)
#' }
#' @export
plot_clusters <- function(x, ...) {
stopifnot_(
is_tna(x),
"Argument {.arg x} must be a {.cls tna} object."
)
result <- list()
matrices <- x$transits
for (clus in seq_along(matrices)) {
plot.tna(x, cluster = clus, ...)
}
}
Plot Centrality Measures}
\usage{
\method{plot}{tna_centralities}(
  x,
  model = NULL,
  reorder = TRUE,
  ncol = 3,
  scales = c("free_x", "fixed"),
  colors = NULL,
  labels = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{tna_centralities}.}

\item{model}{An object of class \code{tna}.}

\item{reorder}{A \code{logical} value indicating whether to reorder the values
for each centrality in a descending order. The default is \code{TRUE}.}

\item{ncol}{Number of columns to use for the facets. The default is 3.}

\item{scales}{Either \code{"fixed"} or \code{"free_x"} (the default). If \code{"free_x"},
the horizontal axis is scaled individually in each facet. If \code{"fixed"},
the same values are used for all axes.}

\item{colors}{The colors for each node (default is the model colors
if the \code{tna} model object is passed, otherwise \code{"black"}).}

\item{labels}{A \code{logical} value indicating whether to show the centrality
numeric values. The default is \code{TRUE}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object displaying the lollipop charts for each centrality
measure.
}
\description{
Plots the centrality measures of a \code{tna_centralities} object as a
lollipop chart. The resulting plot includes facets for each centrality
measure, showing the values for each state. The returned plot is a
\code{ggplot2} object, so it can be easily modified and styled. See
\code{\link[=centralities]{centralities()}} for details on the centrality measures.
}
\examples{
tna_model <- build_tna(engagement)
cm <- centralities(tna_model)
plot(cm)
plot(cm, ncol = 4, reorder = TRUE)

}
\seealso{
Core functions
\code{\link{build_tna}()},
\code{\link{centralities}()},
\code{\link{plot.tna}()},
\code{\link{plot_compare}()}
}
\concept{core}
