% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna_cliques}
\alias{plot.tna_cliques}
\title{Plot Cliques of a TNA Network}
\usage{
\method{plot}{tna_cliques}(x, n = 6, first = 1, show_loops = FALSE, colors, ask = TRUE, ...)
}
\arguments{
\item{x}{A \code{tna_cliques} object.}

\item{n}{An \code{integer} defining the maximum number of cliques to show.
The defaults is \code{6}.}

\item{first}{An \code{integer} giving the index of the first clique to show.
The default index is \code{1}.}

\item{show_loops}{A \code{logical} value indicating whether to include loops
in the plots or not.}

\item{ask}{A \code{logical} value. When \code{TRUE}, show plots one by one and asks
to plot the next plot in interactive mode.}

\item{...}{Ignored.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Plot Cliques of a TNA Network
}
\examples{
model <- tna(group_regulation)
cliq <- cliques(model, size = 2)
plot(cliq, n = 1, ask = FALSE)

}
\seealso{
Clique-related functions
\code{\link{cliques}()},
\code{\link{plot.group_tna_cliques}()},
\code{\link{print.group_tna_cliques}()},
\code{\link{print.tna_cliques}()}
}
\concept{cliques}
