% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna_comparison}
\alias{plot.tna_comparison}
\title{Plot the results of comparing two \code{tna} models or matrices}
\usage{
\method{plot}{tna_comparison}(x, type = "heatmap", population = "diff", name_x = "x", name_y = "y", ...)
}
\arguments{
\item{x}{A \code{tna_comparison} object.}

\item{type}{A \code{character} string naming the type of plot to produce. The
available options are \code{"heatmap"} (the default), \code{"scatterplot"},
\code{"centrality_heatmap"}, and \code{"weight_density"}.}

\item{population}{A \code{"character"} string naming the population for which
to produce the heatmaps, i.e, one of \code{"x"}, \code{"y"}, or \code{"diff"} for the
differences. Ignored for \code{type = "scatterplot"}. Defaults to \code{"diff"}.}

\item{name_x}{An optional \code{character} string to use as the name of the
first population in the plots. The default is \code{"x"}.}

\item{name_y}{An optional \code{character} string to use as the name of the
second population in the plots. The default is \code{"y"}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the results of comparing two \code{tna} models or matrices
}
\examples{
model_x <- tna(group_regulation[1:200, ])
model_y <- tna(group_regulation[1001:1200, ])
comp <- compare(model_x, model_y)
plot(comp)

}
\seealso{
Pattern-finding functions
\code{\link{communities}()},
\code{\link{plot.tna_communities}()}
}
\concept{patterns}
