% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna_data}
\alias{plot.tna_data}
\title{Plot State Frequencies and Time Series data}
\usage{
\method{plot}{tna_data}(
  x,
  selected = NULL,
  ncol = NULL,
  scales = c("free", "free_x", "free_y", "fixed")
)
}
\arguments{
\item{x}{A \code{tna_data} object.}

\item{selected}{A specific individual time series. If empty, all are plotted}

\item{ncol}{Number of columns to use for the facets.}

\item{scales}{Any of \code{"fixed"}, \code{"free_x"}, \code{"free_y"}, \code{"free"} (the default).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot State Frequencies and Time Series data
}
\examples{
ts_data <- data.frame(
  id = gl(10, 100),
  time = rep(1:100,10),
  series = c(
    replicate(
      10,
      stats::arima.sim(list(order = c(2, 1, 0), ar = c(0.5, 0.2)), n = 99)
    )
  )
)

data_ts <- import_ts(ts_data, "id", "series", "time", n_states = 5)
}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{import_data}()},
\code{\link{import_ts}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{prepare_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{print.tna_data}()},
\code{\link{simulate.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
