% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna_sequence_comparison}
\alias{plot.tna_sequence_comparison}
\title{Plot a Sequence Comparison}
\usage{
\method{plot}{tna_sequence_comparison}(x, n = 10, legend = TRUE, cells = FALSE, ...)
}
\arguments{
\item{x}{A \code{tna_sequence_comparison} object.}

\item{n}{An \code{integer} giving the number of patterns to plot.
The default is \code{10}.}

\item{legend}{A \code{logical} value indicating whether to show the color scale
legend. The default is \code{TRUE}.}

\item{cells}{A \code{logical} value indicating whether to display the
numeric values in each cell. The default is \code{FALSE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a Sequence Comparison
}
\examples{
# TODO
}
\seealso{
Model comparison functions
\code{\link{compare}()},
\code{\link{compare.group_tna}()},
\code{\link{compare_sequences}()},
\code{\link{plot.tna_comparison}()},
\code{\link{plot_compare}()},
\code{\link{plot_compare.group_tna}()},
\code{\link{print.tna_comparison}()}
}
\concept{comparison}
