% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_associations}
\alias{plot_associations}
\alias{plot_associations.tna}
\title{Plot an Association Network}
\usage{
plot_associations(x, ...)

\method{plot_associations}{tna}(x, edge_color, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{...}{Additional arguments passed to \code{\link[=plot_model]{plot_model()}}.}

\item{edge_color}{An optional \code{character} vector of colors for the edges.
By default, the colors are specified by the magnitude of the
standardized residual.}
}
\value{
A \code{cograph_network} object.
}
\description{
Plot an Association Network
}
\examples{
model <- ftna(group_regulation)
plot_associations(model)

}
