% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot Transition Networks for All Clusters}
\usage{
plot_clusters(x, ...)
}
\arguments{
\item{x}{A \code{tna} object containing transition matrices for different clusters.}

\item{...}{Additional arguments to be passed to the \code{plot.tna} function.}
}
\value{
A series of plots showing the transition networks for each cluster.
}
\description{
This function plots the transition networks for each cluster in a \code{tna} object.
It iterates through the transition matrices for each cluster and generates
corresponding plots using the \code{plot.tna} function.
}
\examples{
\dontrun{
# Assuming `tna_model` is a tna object containing transition matrices
plot_clusters(tna_model)
}
}
\seealso{
Core functions
\code{\link{build_tna}()},
\code{\link{centralities}()},
\code{\link{plot.centralities}()},
\code{\link{plot.tna}()},
\code{\link{plot_compare}()}
}
\concept{core}
