% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_communities}
\alias{plot_communities}
\title{Plot Communities}
\usage{
plot_communities(x, y, cluster = 1L, community, colors)
}
\arguments{
\item{x}{A \code{tna} object, which represents a transition network analysis
object containing data to be plotted. The object should include a component \code{weights}
that can be iterated over to plot individual communities.}

\item{y}{A \code{communities} object generated by the \code{find_communities} method.
Each community detection method maps nodes or points in \code{x}  to specific
communities.}

\item{cluster}{Index of the cluster for which to produce the plot. Defaults
to the first cluster.}

\item{community}{A \code{character} string naming a community detection method to
use for coloring the plot. This can be one of the following:
\itemize{
\item \code{"walktrap"}: A community detection method using short random walks.
\item \code{"fast_greedy"}: A method based on modularity optimization.
\item \code{"label_prop"}: A method that uses label propagation.
\item \code{"infomap"}: A method that uses information flow to detect communities.
\item \code{"edge_betweenness"}: A method that uses edge betweenness to find
communities.
\item \code{"leading_eigen"}: A method using the leading eigenvector of the
modularity matrix.
\item \code{"spinglass"}: A method based on the spinglass model.
}}

\item{colors}{A \code{character} vector of color values used for visualizing
community assignments.}
}
\description{
This function visualizes the communities detected within a \code{tna} object
based on different community detection algorithms and their corresponding
color mappings.
}
\examples{
\dontrun{
plot_communities(tna_model, community_assignment, "walktrap")
}

}
\seealso{
Pattern-finding functions
\code{\link{find_communities}()}
}
\concept{patterns}
