% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_compare.group_tna}
\alias{plot_compare.group_tna}
\title{Plot the difference network between two clusters}
\usage{
\method{plot_compare}{group_tna}(x, i = 1L, j = 2L, ...)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{i}{An \code{integer} index or the name of the principal cluster as a
\code{character} string.}

\item{j}{An \code{integer} index or the name of the secondary cluster as a
\code{character} string.}

\item{...}{Additional arguments passed to \code{\link[=plot_compare.tna]{plot_compare.tna()}}.}
}
\value{
A \code{qgraph} object displaying the difference network between the
two clusters
}
\description{
Plot the difference network between two clusters
}
\examples{
model <- group_model(engagement_mmm)
plot_compare(model)

}
\seealso{
Other comparison: 
\code{\link{compare}()},
\code{\link{compare.group_tna}()},
\code{\link{plot.tna_comparison}()},
\code{\link{plot_compare}()},
\code{\link{print.tna_comparison}()}
}
\concept{comparison}
