% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_frequencies}
\alias{plot_frequencies}
\alias{plot_frequencies.tna}
\title{Plot the Frequency Distribution of States}
\usage{
plot_frequencies(x, ...)

\method{plot_frequencies}{tna}(x, ...)
}
\arguments{
\item{x}{A \code{tna} object created from sequence data.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the Frequency Distribution of States
}
\examples{
model <- tna(group_regulation)
plot_frequencies(model)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{import_data}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{prepare_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{print.tna_data}()},
\code{\link{simulate.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
