% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mosaic}
\alias{plot_mosaic}
\title{Create a mosaic plot of transitions}
\usage{
plot_mosaic(x, digits = 1)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{digits}{An \code{integer} that determines the number of digits to use
for the chi-square test statistic and the p-value in the plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a mosaic plot of transitions
}
\examples{
ftna_model <- ftna(group_regulation)
plot_mosaic(ftna_model)

}
