% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tna}
\alias{plot_tna}
\title{Plot a Transition Network from a Matrix of Edge Weights}
\usage{
plot_tna(
  x,
  labels,
  colors,
  edge.labels = TRUE,
  layout = "circle",
  mar = rep(5, 4),
  cut = 0.1,
  minimum = 0.05,
  theme = "colorblind",
  ...
)
}
\arguments{
\item{x}{A square \code{matrix} of edge weights.}

\item{labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{colors}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{edge.labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{layout}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{mar}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{cut}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}. Defaults to the smallest 20th percentile
of the weights.}

\item{minimum}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}. Defaults to the smallest 10th
percentile of the weights.}

\item{theme}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{...}{Additional arguments passed to \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}
}
\value{
See \code{\link[=plot.tna]{plot.tna()}}.
}
\description{
Plot a Transition Network from a Matrix of Edge Weights
}
\examples{
m <- matrix(rexp(25), 5, 5)
plot_tna(m)

}
