% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{print.group_model_bootstrap}
\alias{print.group_model_bootstrap}
\title{Print \code{group_model} Bootstrap Results}
\usage{
\method{print}{group_model_bootstrap}(x, ...)
}
\arguments{
\item{x}{A \code{group_model_bootstrap} object.}

\item{...}{Ignored.}

\item{digits}{An \code{integer} giving the minimal number of
\emph{significant} digits to print.}

\item{type}{A \code{character} vector giving the type of edges to print.
The default option \code{"both"} prints both statistically significant and
non-significant edges, \code{"sig"} prints only significant edges, and \code{"nonsig"}
prints only the non-significant edges.}
}
\description{
Print \code{group_model} Bootstrap Results
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_model}()},
\code{\link{deprune.group_model}()},
\code{\link{estimate_cs}()},
\code{\link{group_model}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_model}()},
\code{\link{print.group_model_centralities}()},
\code{\link{print.group_model_communities}()},
\code{\link{print.summary.group_model}()},
\code{\link{print.summary.group_model_bootstrap}()},
\code{\link{prune.group_model}()},
\code{\link{rename_groups}()},
\code{\link{reprune.group_model}()},
\code{\link{summary.group_model}()}
}
\concept{clusters}
