% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{print.group_model_cliques}
\alias{print.group_model_cliques}
\title{Print \code{group_model} Found Cliques}
\usage{
\method{print}{group_model_cliques}(x, ...)
}
\arguments{
\item{x}{A \code{group_model_cliques} object.}

\item{...}{Ignored.}

\item{n}{An \code{integer} defining the maximum number of cliques to show.
The defaults is \code{6}.}

\item{first}{An \code{integer} giving the index of the first clique to show.
The default index is \code{1}.}

\item{digits}{An \code{integer} giving the minimal number of
\emph{significant} digits to print.}
}
\description{
Print \code{group_model} Found Cliques
}
