% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.group_tna}
\alias{print.group_tna}
\title{Print a \code{group_tna} Object}
\usage{
\method{print}{group_tna}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{...}{Arguments passed to \code{\link[=print.tna]{print.tna()}}.}
}
\value{
\code{x} (invisibly).
}
\description{
Print a \code{group_tna} Object
}
\examples{
model <- group_model(engagement_mmm)
print(model)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
