% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{print.group_tna_communities}
\alias{print.group_tna_communities}
\title{Print \code{group_tna} Detected Communities}
\usage{
\method{print}{group_tna_communities}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna_communities} object.}

\item{...}{Ignored.}
}
\description{
Print \code{group_tna} Detected Communities
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{prune.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
