% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{print.summary.group_model}
\alias{print.summary.group_model}
\title{Print the summary of a grouped Transition Network Analysis Model}
\usage{
\method{print}{summary.group_model}(x, ...)
}
\arguments{
\item{x}{A \code{summary.group_model} object.}

\item{...}{Ignored}
}
\description{
Print the summary of a grouped Transition Network Analysis Model
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_model}()},
\code{\link{deprune.group_model}()},
\code{\link{estimate_cs}()},
\code{\link{group_model}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_model}()},
\code{\link{print.group_model_bootstrap}()},
\code{\link{print.group_model_centralities}()},
\code{\link{print.group_model_communities}()},
\code{\link{print.summary.group_model_bootstrap}()},
\code{\link{prune.group_model}()},
\code{\link{rename_groups}()},
\code{\link{reprune.group_model}()},
\code{\link{summary.group_model}()}
}
\concept{clusters}
