% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.tna}
\alias{print.summary.tna}
\title{Print a TNA Summary}
\usage{
\method{print}{summary.tna}(x, ...)
}
\arguments{
\item{x}{A \code{summary.tna} object.}

\item{...}{Ignored.}
}
\value{
A \code{summary.tna} object (invisibly) containing the TNA model network
metrics and values.
}
\description{
Print a TNA Summary
}
\examples{
model <- tna(engagement)
print(summary(model))

}
