% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.tna_bootstrap}
\alias{print.summary.tna_bootstrap}
\title{Print Bootstrap Summary}
\usage{
\method{print}{summary.tna_bootstrap}(x, ...)
}
\arguments{
\item{x}{A \code{summary.tna_bootstrap} object.}

\item{...}{Ignored.}
}
\description{
Print Bootstrap Summary
}
