% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.tna_mmm}
\alias{print.summary.tna_mmm}
\title{Print a Summary of a Mixture Markov Model Fit}
\usage{
\method{print}{summary.tna_mmm}(x, digits = 3L, ...)
}
\arguments{
\item{x}{A \code{summary.tna_mmm} object.}

\item{...}{Arguments passed to the generic \code{print} method.}
}
\value{
A \code{summary.tna_mmm} object (invisibly) with details of the model fit.
}
\description{
Print a Summary of a Mixture Markov Model Fit
}
\examples{
sumr <- summary(engagement_tna_mmm)
print(sumr)

}
\concept{groups}
