% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna}
\alias{print.tna}
\title{Print a \code{tna} object}
\usage{
\method{print}{tna}(x, digits = 3, generic = FALSE, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{digits}{An \code{integer} giving the number of decimal digits to print.
Defaults to \code{3}.}

\item{generic}{A \code{logical} value. If \code{TRUE}, use generic print method
instead. Defaults to \code{FALSE}.}

\item{...}{Ignored.}
}
\description{
Print a \code{tna} object
}
