% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_cliques}
\alias{print.tna_cliques}
\title{Print Found Cliques}
\usage{
\method{print}{tna_cliques}(x, ...)
}
\arguments{
\item{x}{A \code{tna_cliques} object.}

\item{...}{Ignored.}
}
\description{
Print Found Cliques
}
