% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_communities}
\alias{print.tna_communities}
\title{Print Detected Communities}
\usage{
\method{print}{tna_communities}(x, ...)
}
\arguments{
\item{x}{A \code{tna_communities} object.}

\item{...}{Additional arguments passed to the generic \code{print} method.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Detected Communities
}
\examples{
model <- tna(group_regulation)
comm <- communities(model)
print(comm)

}
\seealso{
Community detection functions
\code{\link{communities}()},
\code{\link{plot.group_tna_communities}()},
\code{\link{plot.tna_communities}()},
\code{\link{print.group_tna_communities}()}
}
\concept{communities}
