% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_comparison}
\alias{print.tna_comparison}
\title{Print Comparison Results}
\usage{
\method{print}{tna_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{tna_comparison} object.}

\item{...}{Additional arguments passed to the tibble \code{print} method.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Comparison Results
}
\examples{
model_x <- tna(group_regulation[1:200, ])
model_y <- tna(group_regulation[1001:1200, ])
comp <- compare(model_x, model_y)
print(comp)

}
\seealso{
Model comparison functions
\code{\link{compare}()},
\code{\link{compare.group_tna}()},
\code{\link{compare_sequences}()},
\code{\link{plot.tna_comparison}()},
\code{\link{plot.tna_sequence_comparison}()},
\code{\link{plot_compare}()},
\code{\link{plot_compare.group_tna}()},
\code{\link{print.tna_sequence_comparison}()}
}
\concept{comparison}
