% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_data}
\alias{print.tna_data}
\title{Print a TNA data object}
\usage{
\method{print}{tna_data}(x, format = "wide", ...)
}
\arguments{
\item{x}{A \code{tna_data} object.}

\item{format}{A \code{character} string that defines the format of the printed
tibble. Accepts either \code{"wide"} (default) for wide format data or \code{"long"}
for long format.}

\item{...}{Arguments passed to the tibble \code{print} method.}
}
\value{
\code{x} (invisibly).
}
\description{
Print a TNA data object
}
\examples{
data_single_session <- tibble::tibble(
  action = c(
    "view", "click", "add_cart", "view", "checkout", "view", "click", "share"
   )
)
results_single <- prepare_data(data_single_session, action = "action")
print(results_single, format = "wide")
print(results_single, format = "long")

}
