% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_mmm}
\alias{print.tna_mmm}
\title{Print a Mixture Markov Model Fit}
\usage{
\method{print}{tna_mmm}(x, digits = 3L, ...)
}
\arguments{
\item{x}{A \code{tna_mmm} object.}

\item{...}{Arguments passed to the generic \code{print} method}
}
\value{
\code{x} (invisibly)
}
\description{
Print a Mixture Markov Model Fit
}
\examples{
print(engagement_tna_mmm)

}
\seealso{
Other groups: 
\code{\link{print.summary.tna_mmm}()}
}
\concept{groups}
