% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_reliability}
\alias{print.tna_reliability}
\title{Print Reliability Analysis Results}
\usage{
\method{print}{tna_reliability}(x, summary_metrics, ...)
}
\arguments{
\item{x}{A \code{tna_reliability} object.}

\item{summary_metrics}{A \code{character} vector of metrics to display.}

\item{...}{Arguments passed to the generic \code{print} method.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Reliability Analysis Results
}
\examples{
# Small number of iterations for CRAN
rel <- reliability(engagement, iter = 20)
print(rel)

}
\seealso{
Model comparison functions
\code{\link{compare}()},
\code{\link{compare.group_tna}()},
\code{\link{compare_sequences}()},
\code{\link{plot.tna_comparison}()},
\code{\link{plot.tna_reliability}()},
\code{\link{plot.tna_sequence_comparison}()},
\code{\link{plot_compare}()},
\code{\link{plot_compare.group_tna}()},
\code{\link{print.tna_comparison}()},
\code{\link{print.tna_sequence_comparison}()}
}
\concept{comparison}
