% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_sequence_comparison}
\alias{print.tna_sequence_comparison}
\title{Print a Comparison of Sequences}
\usage{
\method{print}{tna_sequence_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{tna_sequence_comparison} object.}

\item{...}{Arguments passed to the generic \code{print} method.}
}
\value{
\code{x} (invisibly).
}
\description{
Print a Comparison of Sequences
}
\examples{
group <- c(rep("High", 1000), rep("Low", 1000))
comp <- compare_sequences(group_regulation, group)
print(comp)

}
\seealso{
Model comparison functions
\code{\link{compare}()},
\code{\link{compare.group_tna}()},
\code{\link{compare_sequences}()},
\code{\link{plot.tna_comparison}()},
\code{\link{plot.tna_sequence_comparison}()},
\code{\link{plot_compare}()},
\code{\link{plot_compare.group_tna}()},
\code{\link{print.tna_comparison}()}
}
\concept{comparison}
