% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_stability}
\alias{print.tna_stability}
\title{Print Centrality Stability Results}
\usage{
\method{print}{tna_stability}(x, ...)
}
\arguments{
\item{x}{A \code{tna_stability} object.}

\item{...}{Ignored.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Centrality Stability Results
}
\examples{
model <- tna(engagement)
# Small number of iterations and drop proportions for CRAN
cs <- estimate_cs(
  model,
  measures = c("InStrength", "OutStrength"),
  drop_prop = seq(0.3, 0.9, by = 0.2),
  iter = 10
)
print(cs)

}
