% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{prune.group_tna}
\alias{prune.group_tna}
\title{Prune a \code{group_tna} network based on transition probabilities}
\usage{
\method{prune}{group_tna}(x, ...)
}
\arguments{
\item{x}{An object of class \code{group_tna}}

\item{...}{Arguments passed to \code{\link[=bootstrap]{bootstrap()}} when
using \code{method = "bootstrap"} and when a \code{group_tna_bootstrap} is not supplied.}

\item{method}{A \code{character} string describing the pruning method.
The available options are \code{"threshold"}, \code{"lowest"}, \code{"bootstrap"} and
\code{"disparity"}, corresponding to the methods listed in Details. The default
is \code{"threshold"}.}

\item{threshold}{A numeric value specifying the edge weight threshold.
Edges with weights below or equal to this threshold will be considered for
removal.}

\item{lowest}{A \code{numeric} value specifying the lowest percentage
of non-zero edges. This percentage of edges with the lowest weights will be
considered for removal. The default is \code{0.05}.}

\item{level}{A \code{numeric} value representing the significance level for the
disparity filter. Defaults to \code{0.5}.}

\item{boot}{A \code{group_tna_bootstrap} object to be used for pruning with method
\code{"boot"}. The method argument is ignored if this argument is supplied.}
}
\value{
A pruned \code{group_tna} object. Details on the pruning can be viewed with
\code{\link[=pruning_details]{pruning_details()}}. The original model can be restored with \code{\link[=deprune]{deprune()}}.
}
\description{
Prunes a set of networks represented by a \code{group_tna} object by removing
edges based on a specified threshold, lowest percent of non-zero edge
weights, or the disparity filter algorithm (Serrano et al., 2009).
It ensures the networks remain weakly connected.
}
\examples{
group <- c(rep("High",100),rep("Low",100))
model <- group_tna(engagement, group = group)
pruned_threshold <- prune(model, method = "threshold", threshold = 0.1)
pruned_percentile <- prune(model,method = "lowest", lowest = 0.05)
pruned_disparity <- prune(model, method = "disparity", level = 0.5)

}
\references{
Serrano, M. A., Boguna, M., & Vespignani, A. (2009). Extracting the
multiscale backbone of complex weighted networks.
\emph{Proceedings of the National Academy of Sciences, 106},
6483-6488. \doi{10.1073/pnas.0808904106}
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
