% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{prune.tna}
\alias{prune.tna}
\title{Prune a \code{tna} network based on transition probabilities}
\usage{
\method{prune}{tna}(
  x,
  method = "threshold",
  threshold = 0.1,
  lowest = 0.05,
  level = 0.5,
  boot = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{tna}}

\item{method}{A \code{character} string describing the pruning method.
The available options are \code{"threshold"}, \code{"lowest"}, \code{"bootstrap"} and
\code{"disparity"}, corresponding to the methods listed in Details. The default
is \code{"threshold"}.}

\item{threshold}{A numeric value specifying the edge weight threshold.
Edges with weights below or equal to this threshold will be considered for
removal.}

\item{lowest}{A \code{numeric} value specifying the lowest percentage
of non-zero edges. This percentage of edges with the lowest weights will be
considered for removal. The default is \code{0.05}.}

\item{level}{A \code{numeric} value representing the significance level for the
disparity filter. Defaults to \code{0.5}.}

\item{boot}{A \code{tna_bootstrap} object to be used for pruning with method
\code{"boot"}. The method argument is ignored if this argument is supplied.}

\item{...}{Arguments passed to \code{\link[=bootstrap]{bootstrap()}} when
using \code{method = "bootstrap"} and when a \code{tna_bootstrap} is not supplied.}
}
\value{
A pruned \code{tna} object. Details on the pruning can be viewed with
\code{\link[=pruning_details]{pruning_details()}}. The original model can be restored with \code{\link[=deprune]{deprune()}}.
}
\description{
Prunes a network represented by a \code{tna} object by removing
edges based on a specified threshold, lowest percent of non-zero edge
weights, or the disparity filter algorithm (Serrano et al., 2009).
It ensures the network remains weakly connected.
}
\examples{
model <- tna(group_regulation)
pruned_threshold <- prune(model, method = "threshold", threshold = 0.1)
pruned_percentile <- prune(model,method = "lowest", lowest = 0.05)
pruned_disparity <- prune(model, method = "disparity", level = 0.5)

}
\references{
Serrano, M. A., Boguna, M., & Vespignani, A. (2009). Extracting the
multiscale backbone of complex weighted networks.
\emph{Proceedings of the National Academy of Sciences, 106},
6483-6488. \doi{10.1073/pnas.0808904106}
}
\seealso{
Evaluation and validation functions
\code{\link{bootstrap}()},
\code{\link{permutation_test}()},
\code{\link{prune}()},
\code{\link{pruning_details.group_model}()}
}
\concept{evaluation}
