% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R, R/pruning.R
\name{pruning_details.group_tna}
\alias{pruning_details.group_tna}
\alias{pruning_details}
\alias{pruning_details.tna}
\title{Print Detailed Information on the Pruning Results}
\usage{
\method{pruning_details}{group_tna}(x, ...)

pruning_details(x, ...)

\method{pruning_details}{tna}(x, removed_edges = TRUE, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{...}{Ignored.}

\item{removed_edges}{Should a \code{data.frame} of removed edges be printed?
The default is \code{FALSE}.}
}
\description{
Print Detailed Information on the Pruning Results
}
\seealso{
Cluster-related functions
\code{\link{as.igraph.group_tna}()},
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{prune.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
