% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R, R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{simulate}
\alias{as.igraph}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{igraph}{\code{\link[igraph]{as.igraph}}}

  \item{stats}{\code{\link[stats]{simulate}}}
}}

