% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{reprune}
\alias{reprune}
\title{Restore Previous Pruning of a Transition Network Analysis Model}
\usage{
reprune(x, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{...}{Ignored.}
}
\description{
Restore Previous Pruning of a Transition Network Analysis Model
}
