% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.group_tna}
\alias{simulate.group_tna}
\title{Simulate Data from a Group Transition Network Analysis Model}
\usage{
\method{simulate}{group_tna}(
  object,
  nsim = 1,
  seed = NULL,
  max_len = 100L,
  na_range = c(0L, 0L),
  zero_row = c("self", "uniform"),
  format = c("wide", "long"),
  ...
)
}
\arguments{
\item{object}{A \code{group_tna} object. The edge weights must be transition
probabilities, i.e., the model must have \code{type = "relative"} or
\code{type = "absolute"}.}

\item{nsim}{An \code{integer} giving the number of sequences to simulate.
The default is 1. It can also be a vector whose length matches
the number of groups in the model.}

\item{seed}{Ignored. Please use \code{\link[=set.seed]{set.seed()}} manually.}

\item{max_len}{An \code{integer} giving the maximum length of the simulated
sequences. When no missing values are generated, this is the length of
all simulated sequences. It can also be a vector whose length matches
the number of groups in the model.}

\item{na_range}{An \code{integer} vector of length 2 giving the minimum and
maximum number of missing values to generate for each sequence. The number
of missing values is drawn uniformly from this range. If both values are
zero (the default), no missing values are generated.}

\item{format}{A \code{character} indicating whether the data should be in \code{wide} or
\code{long} format.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame} of the simulated sequence data with \code{nsim} rows and
\code{max_len} columns.
}
\description{
Simulate Data from a Group Transition Network Analysis Model
}
\examples{
model <- group_tna(group_regulation,
                   group = c(rep("High",1000), rep("Low",1000)))
sim <- simulate(model, nsim = 10, max_len = 10)

}
\seealso{
Other data: 
\code{\link{import_data}()},
\code{\link{import_onehot}()},
\code{\link{prepare_data}()},
\code{\link{print.tna_data}()},
\code{\link{simulate.tna}()}
}
\concept{data}
