% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{summary.group_tna}
\alias{summary.group_tna}
\title{Summarize a grouped Transition Network Analysis Model}
\usage{
\method{summary}{group_tna}(x, combined = FALSE)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{combined}{A logical indicating whether the summary results should be
combined into a single dataframe for all clusters (defaults to \code{FALSE})}
}
\value{
A \code{data.frame} object if combined is \code{FALSE} or a \code{list} if
combined is \code{TRUE}
}
\description{
Summarize a grouped Transition Network Analysis Model
}
\seealso{
Cluster-related functions
\code{\link{as.igraph.group_tna}()},
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{prune.group_tna}()},
\code{\link{pruning_details.group_tna}()},
\code{\link{reprune.group_tna}()}
}
\concept{clusters}
