% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tna_data}
\alias{summary.tna_data}
\title{Summarize TNA Data}
\usage{
\method{summary}{tna_data}(object, ...)
}
\arguments{
\item{object}{A \code{tna_data} object.}

\item{...}{Arguments passed to \code{\link[=summary.data.frame]{summary.data.frame()}}.}
}
\value{
A \code{summary.tna_data} object, which a \code{list} of summaries for the
long data, sequence data, and metadata.
}
\description{
Summarize TNA Data
}
\examples{
data <- tibble::tibble(
  user = c("A", "A", "A", "B", "B", "C", "C", "C"),
  time = c(
    "2023-01-01 10:00:00", "2023-01-01 10:05:00",
    "2023-01-01 10:20:00", "2023-01-01 12:00:00",
    "2023-01-01 12:02:00", "2023-01-01 14:00:00",
    "2023-01-01 14:05:00", "2023-01-01 14:10:00"
  ),
  action = c(
    "view", "click", "add_cart", "view",
    "checkout", "view", "click", "share"
   )
)
results <- prepare_data(
  data, actor = "user", time = "time", action = "action"
)
summary(result)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{import_data}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{prepare_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.summary.tna_data}()},
\code{\link{print.tna}()},
\code{\link{print.tna_data}()},
\code{\link{simulate.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}},
\code{\link{tna2sts}()}
}
\concept{basic}
