% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tna_mmm}
\alias{summary.tna_mmm}
\title{Summarize a Mixture Markov Model Fit}
\usage{
\method{summary}{tna_mmm}(object, ...)
}
\arguments{
\item{object}{A \code{tna_mmm} object.}

\item{...}{Not used.}
}
\value{
A \code{summary.tna_mmm} object containing the log-likelihood value,
the regression coefficients, the variance-covariance matrix and other
details.
}
\description{
Summarize a Mixture Markov Model Fit
}
\examples{
summary(engagement_tna_mmm)

}
