% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tna2sts}
\alias{tna2sts}
\alias{tna2sts.tna}
\alias{tna2sts.group_tna}
\alias{tna2sts.tna_data}
\title{Convert TNA Sequence Data into STS Format}
\usage{
tna2sts(x, ...)

\method{tna2sts}{tna}(x, void = "\%", nr = "*", ...)

\method{tna2sts}{group_tna}(x, void = "\%", nr = "*", ...)

\method{tna2sts}{tna_data}(x, void = "\%", nr = "*", ...)
}
\arguments{
\item{x}{A \code{tna}, \code{group_tna}, or \code{tna_data} object.}

\item{...}{Ignored.}

\item{void}{The internal code used by \code{TraMineR} for representing void
elements in the sequences. The default is \code{"\%"}.}

\item{nr}{The internal code used by \code{TraMineR} for representing real
missing elements in the sequences. The default is \code{"*"}.}
}
\value{
An \code{stslist} object of the sequence data.
}
\description{
Converts data from the formats used by \code{tna} into \code{stslist} objects used
by, e.g., \code{TraMineR} and \code{seqHMM}. Gaps and left-missing values are
converted to missing values and right-missing values are removed.
}
\examples{
model <- tna(group_regulation)
tna2sts(model)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{import_data}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{prepare_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.summary.tna_data}()},
\code{\link{print.tna}()},
\code{\link{print.tna_data}()},
\code{\link{simulate.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{summary.tna_data}()},
\code{\link{tna-package}}
}
\concept{basic}
