% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.tna_mmm}
\alias{vcov.tna_mmm}
\alias{vcov.summary.tna_mmm}
\title{Variance-Covariance Matrix of the Covariate Coefficients of a Mixture Markov Model}
\usage{
\method{vcov}{tna_mmm}(object, ...)

\method{vcov}{summary.tna_mmm}(object, ...)
}
\arguments{
\item{object}{A \code{tna_mmm} or a \code{summary.tna_mmm} object.}

\item{...}{Not used.}
}
\value{
A \code{matrix} representing the variance-covariance matrix.
}
\description{
Variance-Covariance Matrix of the Covariate Coefficients of a Mixture Markov Model
}
\examples{
vcov(engagement_tna_mmm)

}
