% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.igraph.group_tna}
\alias{as.igraph.group_tna}
\title{Coerce a Specific Group from a \code{group_tna} Object into an \code{igraph} Object.}
\usage{
\method{as.igraph}{group_tna}(x, which, ...)
}
\arguments{
\item{x}{The object to convert.}

\item{which}{The number or name of the group.}

\item{...}{Additional arguments. None currently.}
}
\value{
An \code{igraph} object.
}
\description{
Coerce a Specific Group from a \code{group_tna} Object into an \code{igraph} Object.
}
\seealso{
Helper functions
\code{\link{as.igraph.matrix}()},
\code{\link{as.igraph.tna}()}
}
\concept{helpers}
