% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.igraph.tna}
\alias{as.igraph.tna}
\title{Coerce a \code{tna} Object into an \code{igraph} Object.}
\usage{
\method{as.igraph}{tna}(x, ...)
}
\arguments{
\item{x}{The object to convert.}

\item{...}{Additional arguments. None currently.}
}
\value{
An \code{igraph} object.
}
\description{
Coerce a \code{tna} Object into an \code{igraph} Object.
}
\seealso{
Helper functions
\code{\link{as.igraph.group_tna}()},
\code{\link{as.igraph.matrix}()}
}
\concept{helpers}
