% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_tna}
\alias{build_tna}
\title{Build a Transition Network Analysis Model}
\usage{
build_tna(sequence)
}
\arguments{
\item{sequence}{A sequence of events or states to be used for building the Markov model.}
}
\value{
An object of class \code{tna} which is a list containing the following elements:
\itemize{
\item \code{Matrix} A matrix of transition probabilities.
\item \code{Matrix0} A transition probability matrix without loops (the diagonal is set to zero).
\item \code{inits} A vector of initial state probabilities.
\item \code{Labels} The labels of the states.
\item \code{colors} A color palette for the states.
}
}
\description{
This function constructs a transition network analysis (TNA) model from a given sequence.
It takes a sequence of events or states and builds a Markov model using  \code{\link[seqHMM:build_mm]{seqHMM::build_mm()}}.
It extracts the transition probabilities and initial probabilities from the model and stores them in a list
along with the state labels. Additionally, it creates a transition matrix with zero diagonal entries (without loops).
}
\examples{
data(biofam3c, package = "seqHMM")

# Preparing the sequence data
seq_data <- TraMineR::seqdef(as.data.frame(biofam3c[1:3]))

# Building a transition matrix from the sequence data
tna_model <- build_tna(seq_data)

}
