% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralities.R
\name{centralities}
\alias{centralities}
\alias{centralities.tna}
\alias{centralities.matrix}
\title{Calculate Centralities for a Transition Matrix}
\usage{
centralities(x, ...)

\method{centralities}{tna}(x, loops = TRUE, ...)

\method{centralities}{matrix}(x, ...)
}
\arguments{
\item{x}{A square matrix representing transition probabilities or adjacency,
or a \code{tna} object.}

\item{...}{Ignored.}

\item{loops}{A \code{logical} value indicating whether to include loops in the
network when computing the centrality measures (default is \code{TRUE}).}
}
\value{
A \code{data.frame} containing centrality measures for each interaction.
}
\description{
This function calculates the following centrality measures:
}
\details{
\itemize{
\item \code{OutStrength} Outgoing strength centrality, calculated using \code{\link[igraph:strength]{igraph::strength()}} with \code{mode = "out"}.
It measures the total weight of the outgoing edges from each node.
\item \code{InStrength} Incoming strength centrality, calculated using \code{\link[igraph:strength]{igraph::strength()}} with \code{mode = "in"}.
It measures the total weight of the incoming edges to each node.
\item \code{ClosenessIn} Closeness centrality (incoming), calculated using \code{\link[igraph:closeness]{igraph::closeness()}} with \code{mode = "in"}.
It measures how close a node is to all other nodes based on the incoming paths.
\item \code{ClosenessOut} Closeness centrality (outgoing), calculated using \code{\link[igraph:closeness]{igraph::closeness()}} with \code{mode = "out"}.
It measures how close a node is to all other nodes based on the outgoing paths.
\item \code{Closeness} Closeness centrality (overall), calculated using \code{\link[igraph:closeness]{igraph::closeness()}} with \code{mode = "all"}.
It measures how close a node is to all other nodes based on both incoming and outgoing paths.
\item \code{Betweenness} Betweenness centrality based on randomized shortest paths (Kivimäki et al. 2014).
It measures the extent to which a node lies on the shortest paths between other nodes.
\item \code{Diffusion} Diffusion centrality of Banerjee et.al. (2014).
It measures the influence of a node in spreading information through the network.
\item \code{Clustering} Signed clustering coefficient of Zhang and Horvath (2005)
based on the symmetric adjacency matrix (sum of the adjacency matrix and its transpose).
It measures the degree to which nodes tend to cluster together.
}
}
\examples{
tna_model <- build_tna(engagement)

# Centrality measures including loops in the network
centralities(tna_model)

# Centrality measures excluding loops in the network
centralities(tna_model, loops = FALSE)

}
\references{
Banerjee, A., A. Chandrasekhar, E. Duflo, and M. Jackson (2014).
Gossip: Identifying Central Individuals in a Social Network.
Working Paper.

Kivimaki, I., Lebichot, B., Saramaki, J., & Saerens, M. (2016).
Two betweenness centrality measures based on Randomized Shortest Paths.
Scientific Reports, 6, 19668.

Zhang, B., & Horvath, S. (2005).
A general framework for weighted gene co-expression network analysis.
Statistical Applications in Genetics and Molecular Biology, 4(1).
}
