% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques.R
\name{cliques}
\alias{cliques}
\title{Identify Cliques in a Transition Network}
\usage{
cliques(x, cluster = 1, size = 3, threshold = 0)
}
\arguments{
\item{x}{An object of type \code{tna}}

\item{cluster}{An integer specifying which cluster to analyze.
Defaults to \code{1}.}

\item{size}{An \code{integer} specifying the size of the cliques to identify.
Defaults to \code{3} (triads).}

\item{threshold}{A \code{numeric} value that sets the minimum edge weight
for an edge to be considered in the clique. Edges below this value
are ignored. Defaults to \code{0}.}
}
\value{
A \code{cliques} object which is a \code{list} of two elements:
\itemize{
\item \code{weights} is a \code{matrix} of the edge weights in the clique.
\item \code{inits} is a \code{numeric} vector of initial weights for the clique.
}
}
\description{
This function identifies cliques of a specified size in a transition network.
It searches for cliques—complete subgraphs where every pair of nodes is
connected—of size \code{n} in the transition matrix for the specified cluster
in the \code{tna} object.
}
\examples{
\dontrun{
# Find  3-cliques (triads) in the first cluster
cliques_result <- cliques(my_tna_object, cluster = 1, size = 3)

# Find 4-cliques in the second cluster
cliques_result <- cliques(my_tna_object, cluster = 2, size = 4)
}

}
