% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm.R
\name{cluster_mmm}
\alias{cluster_mmm}
\title{Mixture Markov Model Clustering for Sequences}
\usage{
cluster_mmm(
  data,
  cols = seq(1L, ncol(data)),
  formula,
  k,
  cluster_names,
  criterion = "bic",
  progressbar = TRUE,
  parallel = FALSE,
  n_cores,
  cl,
  control
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} where rows represent sequences and
columns represent time points. Missing values (\code{NA}) are handled
appropriately. The columns should be either \code{integer}, \code{character} or
\code{factor} variables.}

\item{cols}{An \code{integer}/\code{character} vector giving the indices/names of the
columns that should be considered as sequence data.
Defaults to all columns, i.e., \code{seq(1, ncol(data))}. Column names not
found in \code{data} will be ignored without warning.}

\item{formula}{An optional \code{formula} object that specifies the covariates
to use for the cluster membership probabilities.}

\item{k}{An \code{integer} vector specifying the numbers of mixture components
(clusters) to fit. The values must be between 2 and the number of
sequences minus 1.}

\item{cluster_names}{An optional \code{character} vector of cluster names.
The default naming scheme is Cluster 1, Cluster 2, etc. The length of
the vector must be the same as the largest value of \code{k}.}

\item{criterion}{A \code{character} string specifying the information criterion
to use for model selection. Either \code{"bic"} (default) for Bayesian
information criterion or \code{"aic"} for Akaike information criterion.}

\item{progressbar}{A \code{logical} value. If \code{TRUE}, a progress bar is displayed.
The default is \code{FALSE}. Disabled when using parallel processing.}

\item{parallel}{A \code{logical} value.}

\item{n_cores}{An \code{integer} specifying the number of cores to use for
parallel processing. The default is 1 for no parallel processing.}

\item{cl}{An optional prespecified cluster object with a registered
parallel backend. Ignores \code{parallel} if provided.}

\item{control}{An optional \code{list} of arguments for the EM algorithm.
The possible arguments are:
\itemize{
\item \code{maxiter}: An \code{integer} for the maximum number of EM iterations per
restart. The default is \code{500}.
\item \code{maxiter_m}: An \code{integer} for the maximum number of Newton-Rhapson
iterations in the M-step (only with covariates). The default is \code{500}.
\item \code{reltol}: A \code{numeric} value specifying the relative convergence
tolerance for the log-likelihood change between iterations.
The default is \code{1e-10}.
\item \code{reltol_m}: A \code{numeric} value specifying the relative tolerance for the
Newton-Rhapson optimization in the M-step (only with covariates).
The default is \code{1e-6}.
\item \code{restarts}: An \code{integer} specifying the number of random restarts to
perform. Multiple restarts help avoid local optima. The default is \code{10}.
\item \code{seed}: An \code{integer} specifying the base random seed.
The initial values are generated for each restart using \code{seed + i} as
the seed where \code{i} is the index of the restart.
\item \code{step}: An \code{integer} defining the step size for the Newton-Rhapson
iterations. The default is \code{1.0}.
}}
}
\description{
Fits a mixture of first-order Markov models to sequence data using the
Expectation-Maximization (EM) algorithm. Also performs model selection
for the number of clusters using information criteria (AIC or BIC).
}
\details{
The function implements a mixture of first-order Markov models where each
component models sequences with initial state probabilities and transition
probabilities between states, along with mixing weights for component
membership. The estimation is carried out using the EM algorithm.

Multiple random restarts are performed to avoid local optima, with the
best solution (highest log-likelihood) selected as the final result.

If \code{k} is a vector, then the model is fitted for each value in \code{k} and
the optimal number of clusters is selected based on the specified
information criterion.
}
\examples{
\dontrun{
# Fit a MMM with k = 3
model1 <- cluster_mmm(engagement, k = 3)

# Find optimal k for k = 2,3,4 using BIC
model2 <- cluster_mmm(engagement, k = 2:4, criterion = "bic")
}

}
\seealso{
Cluster-related functions
\code{\link{communities}()},
\code{\link{group_model}()},
\code{\link{mmm_stats}()},
\code{\link{print.tna_mmm}()},
\code{\link{rename_groups}()}
}
\concept{clusters}
