% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{communities}
\alias{communities}
\alias{communities.tna}
\title{Community Detection for Transition Networks}
\usage{
communities(x, ...)

\method{communities}{tna}(x, gamma = 1, ...)
}
\arguments{
\item{x}{A \code{tna} object that contains transition matrices.}

\item{...}{Ignored.}

\item{gamma}{A numeric parameter that affects the behavior of certain
algorithms like the Spin Glass method. Defaults to \code{1}.}
}
\value{
An object of class \code{tna_communities} which is a \code{list} with an
element for each cluster containing:
\itemize{
\item \code{counts}: A \code{list} with the number of communities found
by each algorithm.
\item \code{assignments}: A \code{data.frame} where each row corresponds to a
node and each column to a community detection algorithm,
with color-coded community assignments.
}
}
\description{
This function detects communities within the transition networks
(represented by the \code{tna} object).
It uses various algorithms to find communities in the graph representation
of transitions and returns a \code{list} of communities for each cluster or a
specified cluster. If multiple transition matrices exist, the function
iterates over each cluster in the \code{tna} object to find communities using
different algorithms. The function uses the \code{igraph} package to convert
the transition matrices into graphs and then applies community detection
algorithms (e.g., Walktrap, Fast Greedy, Label Propagation, Infomap,
Edge Betweenness, Leading Eigenvector, and Spin Glass).
}
\examples{
model <- tna(engagement)
comm <- communities(tna_model)

}
\seealso{
Pattern-finding functions
\code{\link{plot.tna_communities}()}
}
\concept{patterns}
