% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{deprune}
\alias{deprune}
\alias{deprune.tna}
\alias{reprune.tna}
\title{Restore a Pruned Transition Network Analysis Model}
\usage{
deprune(x, ...)

\method{deprune}{tna}(x, ...)

\method{reprune}{tna}(x, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{...}{Ignored.}
}
\description{
Restore a Pruned Transition Network Analysis Model
}
