% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{engagement}
\alias{engagement}
\title{Example data on student engagement}
\format{
A \code{stslist} object (sequence data).
}
\source{
\doi{10.1016/j.compedu.2023.104787}
}
\usage{
engagement
}
\description{
Students' engagement states (Active / Average / Disengaged)
throughout a whole study program. The data was generated synthetically
based on the article "The longitudinal association between engagement and
achievement varies by time, students’ profiles, and achievement state:
A full program study"
}
\seealso{
Other examples: 
\code{\link{engagement_mmm}},
\code{\link{group_regulation}}
}
\concept{examples}
\keyword{datasets}
