% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{engagement}
\alias{engagement}
\title{Example data on student engagement}
\format{
An object of class \code{stslist} (inherits from \code{data.frame}) with 200 rows and 20 columns.
}
\usage{
engagement
}
\description{
Students' engagement states (Active / Average / Disengaged)
throuhgout a whole study program. The data was generated synthetically
based on the article "The longitudinal association between engagement and
achievement varies by time, students’ profiles, and achievement state:
A full program study" (doi: 10.1016/j.compedu.2023.104787)
}
\keyword{datasets}
