% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{engagement_mmm}
\alias{engagement_mmm}
\title{Example mixed Markov model fitted to the \code{engagement} data}
\format{
A \code{mhmm} object.
}
\source{
The data was generated via \code{mixed_markov_model.R} in
\url{https://github.com/sonsoleslp/tna/tree/main/data-raw/}
}
\usage{
engagement_mmm
}
\description{
Example mixed Markov model fitted to the \code{engagement} data
}
\seealso{
Other examples: 
\code{\link{engagement}},
\code{\link{group_regulation}}
}
\concept{examples}
\keyword{datasets}
