% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{data}
\name{group_regulation}
\alias{group_regulation}
\title{Example data on group regulation}
\format{
A \code{data.frame} object.
}
\source{
The data was generated synthetically.
}
\usage{
group_regulation
}
\description{
Students' regulation during collaborative learning. Students' interactions
were coded as:  "adapt", "cohesion", "consensus", "coregulate", "discuss",
"emotion", "monitor", "plan", "synthesis"
}
\seealso{
Other examples: 
\code{\link{engagement}},
\code{\link{engagement_mmm}}
}
\concept{examples}
\keyword{datasets}
