% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{group_tna}
\alias{group_tna}
\alias{group_tna.default}
\alias{group_tna.mhmm}
\title{Build a grouped Transition Network Analysis Model}
\usage{
group_tna(x, ...)

\method{group_tna}{default}(x, group, ...)

\method{group_tna}{mhmm}(x, ...)
}
\arguments{
\item{x}{An \code{mhmm} object from \code{seqHMM}}

\item{...}{Same as \code{tna}}

\item{group}{A vector indicating the cluster assignment of each
row of the data / sequence}
}
\value{
A \code{group_tna} object

A \code{group_tna} object

A \code{group_tna} object
}
\description{
Build a grouped Transition Network Analysis Model

Build a grouped Transition Network Analysis Model providing group/cluster assignments

Build a grouped Transition Network Analysis Model providing an MMM model from \code{seqHMM}
}
\seealso{
Cluster-related functions
\code{\link{as.igraph.group_tna}()},
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{prune.group_tna}()},
\code{\link{pruning_details.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}

Cluster-related functions
\code{\link{as.igraph.group_tna}()},
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{prune.group_tna}()},
\code{\link{pruning_details.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}

Cluster-related functions
\code{\link{as.igraph.group_tna}()},
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{prune.group_tna}()},
\code{\link{pruning_details.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
