% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.group_tna}
\alias{hist.group_tna}
\title{Plot a Histogram of Edge Weights for a \code{group_tna} Object.}
\usage{
\method{hist}{group_tna}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{...}{Additional arguments passed to \code{\link[graphics:hist]{graphics::hist()}}.}
}
\value{
A \code{list} (invisibly) of \code{histogram} objects of the edge weights of
each cluster.
}
\description{
Plot a Histogram of Edge Weights for a \code{group_tna} Object.
}
\examples{
model <- group_model(engagement_mmm)
hist(model)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.tna}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
