% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{mmm_stats}
\alias{mmm_stats}
\title{Retrieve statistics from a mixture Markov model (MMM)}
\usage{
mmm_stats(x, use_t_dist = TRUE, conf_level = 0.95)
}
\arguments{
\item{x}{An \code{mhmm} object}
}
\value{
A \code{data.frame} object
}
\description{
Retrieve statistics from a mixture Markov model (MMM)
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{plot.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{prune.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
