% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm.R
\name{mmm_stats}
\alias{mmm_stats}
\alias{mmm_stats.mhmm}
\title{Retrieve Statistics from a Mixture Markov Model (MMM)}
\usage{
mmm_stats(x, level = 0.05)

\method{mmm_stats}{mhmm}(x, level = 0.05)
}
\arguments{
\item{x}{A \code{mhmm} object.}

\item{level}{A \code{numeric} value representing the significance level for
hypothesis testing and confidence intervals. Defaults to \code{0.05}.}
}
\value{
A \code{data.frame} object.
}
\description{
Retrieve Statistics from a Mixture Markov Model (MMM)
}
\examples{
mmm_stats(engagement_mmm)

}
\seealso{
Cluster-related functions
\code{\link{communities}()},
\code{\link{group_model}()},
\code{\link{rename_groups}()}
}
\concept{clusters}
