% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.centralities}
\alias{plot.centralities}
\title{Plot Centralities for a Transition Matrix}
\usage{
\method{plot}{centralities}(x, ...)
}
\arguments{
\item{x}{An object of class \code{centralities}.}
}
\value{
A \code{ggplot} object displaying the lollipop charts for each centrality measure.
}
\description{
This function calculates several centrality measures using the \code{centralities} method
and then plots these measures using \code{\link[ggcharts:lollipop_chart]{ggcharts::lollipop_chart()}}.
The resulting plot includes facets for each centrality measure, showing the values for each interaction.
The returned plot is a \code{ggplot2} object, so it can be easily modified and styled.
See \code{\link[=centralities.tna]{centralities.tna()}} for details on the centrality measures.
}
\examples{
\dontrun{
  library(TraMineR)
  data(biofam3c)

  # Preparing the sequence data
  seq_data <- seqdef(biofam3c$biofam)

  # Building a transition matrix from the sequence data
  tna_model <- build.tna(seq_data)
  transition_matrix <- tna_model$Matrix

  # Plotting the centralities
  plot.centralities(transition_matrix)
}

}
