% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.centralities}
\alias{plot.centralities}
\title{Plot Centrality Measures}
\usage{
\method{plot}{centralities}(x, ncol = 3, scales = "free", ...)
}
\arguments{
\item{x}{An object of class \code{centralities}.}

\item{ncol}{Number of columns to use for the facets.}

\item{scales}{Either \code{"fixed"} or \code{"free"} (the default). If \code{"free"}, the
horizontal axis is scaled individually in each facet. If \code{"fixed"}, the
same values are used for all axes.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object displaying the lollipop charts for each centrality
measure.
}
\description{
Plots the centrality measures of a \code{centralities} object as a lollipop chart.
The resulting plot includes facets for each centrality measure, showing the
values for each state. The returned plot is a \code{ggplot2} object, so it can be
easily modified and styled. See \code{\link[=centralities]{centralities()}} for details on the
centrality measures.
}
\examples{
tna_model <- build_tna(engagement)
cm <- centralities(tna_model)
plot(cm)

}
