% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.centralities}
\alias{plot.centralities}
\title{Plot Centralities for a Transition Matrix}
\usage{
\method{plot}{centralities}(x, ncol = 3, scales = "free", ...)
}
\arguments{
\item{x}{An object of class \code{centralities}.}

\item{ncol}{Number of columns to use for the facets.}

\item{scales}{Either \code{"fixed"} or \code{"free"} to be used for the x-axis of the
facet scales.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object displaying the lollipop charts for each centrality
measure.
}
\description{
Calculates several centrality measures using the \code{\link[=centralities]{centralities()}} method
and then plots these measures as a lollipop chart. The resulting plot
includes facets for each centrality measure, showing the values for each
interaction. The returned plot is a \code{ggplot2} object, so it can be easily
modified and styled. See \code{\link[=centralities]{centralities()}} for details on the centrality
measures.
}
\examples{
tna_model <- build_tna(engagement)
cm <- centralities(tna_model)
plot(cm)

}
