% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.group_tna}
\alias{plot.group_tna}
\title{Plot a Grouped Transition Network Analysis Model}
\usage{
\method{plot}{group_tna}(x, title, which, ...)
}
\arguments{
\item{x}{A \code{group_model} object.}

\item{title}{A title for each plot. It can be a single string (the same one
will be used for all plots) or a list (one per group)}

\item{which}{An optional \code{integer} vector of groups to plot. By default, all
groups are plotted.}

\item{...}{
  Arguments passed on to \code{\link[=plot.tna]{plot.tna}}
  \describe{
    \item{\code{node_list}}{An optional \code{list} of two \code{character} vectors that define
two mutually exclusive groups of node labels.}
    \item{\code{use_list_order}}{A \code{logical} value. If \code{node_list} is provided,
defines how the order of the nodes in the plot is defined. A \code{TRUE} value
uses the order in \code{node_list}. Otherwise, the nodes are ranked based on
edge weights and ordered according to the rank.}
  }}
}
\value{
\code{NULL} (invisibly).
}
\description{
Plots the transition network of each cluster.
}
\examples{
model <- group_model(engagement_mmm)
plot(model, which = 1)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
