% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.group_tna_cliques}
\alias{plot.group_tna_cliques}
\title{Plot Found Cliques}
\usage{
\method{plot}{group_tna_cliques}(x, title, ...)
}
\arguments{
\item{x}{A \code{group_tna_cliques} object.}

\item{title}{A \code{character} vector of titles to use for each plot.}

\item{...}{Arguments passed to \code{\link[=plot.tna_cliques]{plot.tna_cliques()}}.}
}
\value{
A \code{list} (invisibly) with one element per cluster. Each element
contains a \code{qgraph} plot when only one clique is present per cluster,
otherwise the element is \code{NULL}.
}
\description{
Plot Found Cliques
}
\examples{
model <- group_model(engagement_mmm)
cliq <- cliques(model, size = 2)
plot(cliq, ask = F)

}
\seealso{
Cluster-related functions
\code{\link{bootstrap}()},
\code{\link{centralities}()},
\code{\link{cliques}()},
\code{\link{communities}()},
\code{\link{deprune}()},
\code{\link{estimate_cs}()},
\code{\link{group_model}()},
\code{\link{hist.group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.group_tna_centralities}()},
\code{\link{plot.group_tna_communities}()},
\code{\link{plot.group_tna_stability}()},
\code{\link{plot_compare.group_tna}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_cliques}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.group_tna_stability}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{prune}()},
\code{\link{pruning_details}()},
\code{\link{rename_groups}()},
\code{\link{reprune}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.group_tna_bootstrap}()}
}
\concept{clusters}
