% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna}
\alias{plot.tna}
\title{Plot a Transition Network Analysis Model}
\usage{
\method{plot}{tna}(
  x,
  node_list,
  use_list_order = TRUE,
  scale_nodes,
  scaling_factor = 0.5,
  ...
)
}
\arguments{
\item{x}{A \code{tna} object from \code{\link[=build_model]{build_model()}}.}

\item{node_list}{An optional \code{list} of two \code{character} vectors that define
two mutually exclusive groups of node labels.}

\item{use_list_order}{A \code{logical} value. If \code{node_list} is provided,
defines how the order of the nodes in the plot is defined. A \code{TRUE} value
uses the order in \code{node_list}. Otherwise, the nodes are ranked based on
edge weights and ordered according to the rank.}

\item{scale_nodes}{A \code{character} string giving the name of a centrality
measure to scale the node size by. See \code{\link[=centralities]{centralities()}} for valid names.
If missing (the default), uses default \code{\link[cograph:splot]{cograph::splot()}} scaling.
The value of \code{node_size} provided via \code{...} is used as baseline size.}

\item{scaling_factor}{A \code{numeric} value specifying how strongly to scale
the nodes when \code{scale_nodes} is provided. Values
between 0 and 1 will result in smaller differences and values larger
than 1 will result in greater differences. The default is \code{0.5}.}

\item{...}{Additional arguments passed to \code{\link[cograph:splot]{cograph::splot()}} or
\code{\link[cograph:plot_htna]{cograph::plot_htna()}}.}
}
\value{
A \code{cograph_network} plot of the transition network.
}
\description{
This function plots a transition network analysis (TNA) model using
the \code{cograph} package. The nodes in the graph represent states, with node
sizes corresponding to initial state probabilities. Edge labels represent
the edge weights of the network. See \code{\link[cograph:splot]{cograph::splot()}} for details on
how to further configure the plot.
}
\examples{
model <- tna(group_regulation)
plot(model)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{plot.group_tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
