% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna}
\alias{plot.tna}
\title{Plot a Transition Network Analysis Model}
\usage{
\method{plot}{tna}(
  x,
  pie = x$inits,
  labels = x$labels,
  color = x$colors,
  edge.labels = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{tna} object from \code{\link[=build_tna]{build_tna()}}.}

\item{pie}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{color}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{edge.labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{...}{Additional arguments passed to \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}
}
\value{
A \code{ggplot} of the transition network.
}
\description{
This function plots a transition network analysis (TNA) model using the \code{qgraph} package.
The nodes in the graph represent states, with node sizes corresponding to initial state probabilities.
Edges between nodes represent transition probabilities, and the graph is colored according to the state colors specified in the model.
}
\examples{
\dontrun{
  library("TraMineR")
  data(biofam3c, package = "seqHMM")

  # Preparing the sequence data
  seq_data <- seqdef(biofam3c)

  # Building a transition matrix from the sequence data
  tna_model <- build_tna(seq_data)
  plot(tna_model)
}

}
