% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna}
\alias{plot.tna}
\title{Plot a Transition Network Analysis Model}
\usage{
\method{plot}{tna}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{tna} object from \code{\link[=build_tna]{build_tna()}}.}

\item{digits}{An \code{integer} defining how many digits to show for transition
probabilities}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} of the transition network.
}
\description{
This function plots a transition network analysis (TNA) model using
the \code{qgraph} package. The nodes in the graph represent states, with node
sizes corresponding to initial state probabilities. Edges between nodes
represent the transition probabilities.
}
\examples{
tna_model <- build_tna(engagement)
plot(tna_model)

}
