% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna}
\alias{plot.tna}
\title{Plot Transition Network Analysis Model}
\usage{
plot.tna(model)
}
\arguments{
\item{model}{A list containing the TNA model components, typically the output of the \code{\link{build.tna}} function.}

\item{...}{Additional arguments passed to the \code{qgraph} function.}
}
\value{
A plot of the transition network.
}
\description{
This function plots a transition network analysis (TNA) model using the \code{qgraph} package.
}
\details{
This function uses the \code{qgraph} package to visualize the transition probabilities in the TNA model. The nodes in the graph
represent states, with node sizes corresponding to initial state probabilities. Edges between nodes represent transition
probabilities, and the graph is colored according to the state colors specified in the model.
}
\examples{
\dontrun{
  library(TraMineR)
  data(biofam3c)

  # Preparing the sequence data
  seq_data <- seqdef(biofam3c$biofam)

  # Building a transition matrix from the sequence data
  tna_model <- build.tna(seq_data)
  plot.tna(tna_model)
}

}
