% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna}
\alias{plot.tna}
\title{Plot a Transition Network Analysis Model}
\usage{
\method{plot}{tna}(
  x,
  labels,
  colors,
  pie,
  edge.labels = TRUE,
  layout = "circle",
  mar = rep(5, 4),
  cut,
  minimum,
  theme = "colorblind",
  ...
)
}
\arguments{
\item{x}{A \code{tna} object from \code{\link[=tna]{tna()}}.}

\item{labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{colors}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{pie}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{edge.labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{layout}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{mar}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{cut}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}. Defaults to the smallest 20th percentile
of the weights.}

\item{minimum}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}. Defaults to the smallest 10th
percentile of the weights.}

\item{theme}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{...}{Additional arguments passed to \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}
}
\value{
A \code{qgraph} plot of the transition network.
}
\description{
This function plots a transition network analysis (TNA) model using
the \code{qgraph} package. The nodes in the graph represent states, with node
sizes corresponding to initial state probabilities. Edge labels represent
the edge weights of the network.
}
\examples{
tna_model <- tna(engagement)
plot(tna_model)

}
\seealso{
Core functions
\code{\link{centralities}()},
\code{\link{plot.tna_centralities}()},
\code{\link{plot_compare}()},
\code{\link{tna}()}
}
\concept{core}
