% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna}
\alias{plot.tna}
\title{Plot a Transition Network Analysis Model}
\usage{
\method{plot}{tna}(x, node_list, use_list_order = TRUE, ...)
}
\arguments{
\item{x}{A \code{tna} object from \code{\link[=build_model]{build_model()}}.}

\item{node_list}{An optional \code{list} of two \code{character} vectors that define
two mutually exclusive groups of node labels.}

\item{use_list_order}{A \code{logical} value. If \code{node_list} is provided,
defines how the order of the nodes in the plot is defined. A \code{TRUE} value
uses the order in \code{node_list}. Otherwise, the nodes are ranked based on
edge weights and ordered according to the rank.}

\item{...}{Additional arguments passed to \code{\link[Sonnet:plot_tna]{Sonnet::plot_tna()}} or
\code{\link[Sonnet:plot_htna]{Sonnet::plot_htna()}}.}
}
\value{
TODO
}
\description{
This function plots a transition network analysis (TNA) model.
The nodes in the graph represent states, with node
sizes corresponding to initial state probabilities. Edge labels represent
the edge weights of the network.
}
\examples{
model <- tna(group_regulation)
plot(model)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{plot.group_tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
