% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna}
\alias{plot.tna}
\title{Plot a Transition Network Analysis Model}
\usage{
\method{plot}{tna}(
  x,
  cluster = 1,
  cluster2,
  color = x$colors,
  edge.labels = TRUE,
  labels = x$labels,
  layout = "circle",
  mar = rep(5, 4),
  pie = x$inits[[cluster]],
  cut = 0.1,
  minimum = 0.05,
  theme = "colorblind",
  ...
)
}
\arguments{
\item{x}{A \code{tna} object from \code{\link[=build_tna]{build_tna()}}.}

\item{cluster}{Index of the primary cluster to visualize.
Defaults to the first cluster.}

\item{cluster2}{Optional index of the secondary cluster. If specified,
The difference between the transition probabilities of \code{cluster} and
\code{cluster2} will be plotted.}

\item{color}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{edge.labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{labels}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{layout}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{mar}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{pie}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{cut}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{minimum}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{theme}{See \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}

\item{...}{Additional arguments passed to \code{\link[qgraph:qgraph]{qgraph::qgraph()}}.}
}
\value{
A \code{qgraph} plot of the transition network.
}
\description{
This function plots a transition network analysis (TNA) model using
the \code{qgraph} package. The nodes in the graph represent states, with node
sizes corresponding to initial state probabilities. Edges between nodes
represent the transition probabilities.
}
\examples{
tna_model <- build_tna(engagement)
plot(tna_model)

}
