% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{plot.tna_data}
\alias{plot.tna_data}
\title{Plot State Frequencies and Time Series data}
\usage{
\method{plot}{tna_data}(
  x,
  selected,
  overlay = "v",
  points = FALSE,
  ncol = NULL,
  max_series = 10,
  trend = FALSE,
  scales = c("free", "free_x", "free_y", "fixed")
)
}
\arguments{
\item{x}{A \code{tna_data} object.}

\item{selected}{A \code{character} vector of specific individual time series to
plot. If not provided (the default), all are plotted up to \code{max_series}}

\item{overlay}{A \code{character} string that describes how to plot the overlay
indicating the state assigned to each time point. Can be either \code{"h"} for
a horizontal overlay or \code{"v"} for a vertical overlay (the default). If
\code{NULL}, no overlay is plotted.}

\item{points}{A \code{logical} value indicating whether to plot a point for
each observation colored according to the assigned state.
Defaults to \code{FALSE}.}

\item{ncol}{An \code{integer} giving the number of columns to use for the facets.}

\item{max_series}{An \code{integer} that defines the maximum number
of time-series to plot. The default is 10.}

\item{trend}{A \code{logical} value for whether to add a trend line or not.
The default is \code{FALSE} for no trend line.}

\item{scales}{Any of \code{"fixed"}, \code{"free_x"}, \code{"free_y"}, or
\code{"free"} (the default).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot State Frequencies and Time Series data
}
\examples{
ts_data <- data.frame(
  id = gl(10, 100),
  time = rep(1:100,10),
  series = c(
    replicate(
      10,
      stats::arima.sim(list(order = c(2, 1, 0), ar = c(0.5, 0.2)), n = 99)
    )
  )
)

data_ts <- prepare_ts(ts_data, "id", "series", "time", n_states = 5)
plot(data_ts)

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.tna}()},
\code{\link{print.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
