% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna_permutation}
\alias{plot.tna_permutation}
\title{Plot the Significant Differences from a Permutation Test}
\usage{
\method{plot}{tna_permutation}(x, ...)
}
\arguments{
\item{x}{A \code{tna_permutation} object.}

\item{...}{Arguments passed to \code{\link[=plot_model]{plot_model()}}.}
}
\description{
Plot the Significant Differences from a Permutation Test
}
\examples{
model_x <- tna(group_regulation[1:1000, ])
model_y <- tna(group_regulation[1001:2000, ])
# Small number of iterations for CRAN
perm <- permutation_test(model_x, model_y, iter = 50)
plot(perm)

}
