% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tna_reliability}
\alias{plot.tna_reliability}
\title{Plot Reliability Analysis Results}
\usage{
\method{plot}{tna_reliability}(x, type = "histogram", metric = "Median Abs. Diff.")
}
\arguments{
\item{x}{A \code{tna_reliability} object.}

\item{type}{A \code{character} string specifying the plot type. The options are:
\code{"histogram"} (default), \code{"density"}, or \code{"boxplot"}.}

\item{plot_metric}{A \code{character} string specifying the metric to plot.
The default is the median absolute difference (\code{"Median Abs. Diff."}).}

\item{...}{Ignored}
}
\description{
Plot Reliability Analysis Results
}
\examples{
# Small number of iterations for CRAN
rel <- reliability(engagement, iter = 20)
plot(rel)

}
\seealso{
Model comparison functions
\code{\link{compare}()},
\code{\link{compare.group_tna}()},
\code{\link{compare_sequences}()},
\code{\link{plot.tna_comparison}()},
\code{\link{plot.tna_sequence_comparison}()},
\code{\link{plot_compare}()},
\code{\link{plot_compare.group_tna}()},
\code{\link{print.tna_comparison}()},
\code{\link{print.tna_reliability}()},
\code{\link{print.tna_sequence_comparison}()}
}
\concept{comparison}
