% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{plot_communities}
\alias{plot_communities}
\title{Plot Communities}
\usage{
plot_communities(x, community_assignment, community)
}
\arguments{
\item{x}{A \code{tna} object, which represents a temporal network analysis object containing data to be plotted. The object should include a component \code{transits} that can be iterated over to plot individual communities.}

\item{community_assignment}{A list containing the community assignments generated by the \code{find_communities} function. Each community detection method maps nodes or points in the \code{tna} object to specific communities.}

\item{community}{A named community detection method to use for coloring the plot. This can be one of the following:
\itemize{
\item \code{walktrap}: A community detection method using short random walks.
\item \code{fast_greedy}: A method based on modularity optimization.
\item \code{label_prop}: A method that uses label propagation.
\item \code{infomap}: A method that uses information flow to detect communities.
\item \code{edge_betweenness}: A method that uses edge betweenness to find communities.
\item \code{leading_eigen}: A method using the leading eigenvector of the modularity matrix.
\item \code{spinglass}: A method based on the spinglass model.
}}
}
\description{
This function visualizes the communities detected within a \code{tna} object based on different community detection algorithms and their corresponding color mappings.
}
\details{
The function iterates over the \code{transits} element of the \code{tna} object and generates plots for each using the specified community detection method. The \code{community_assignment} provides the mapping of nodes to communities, and colors are assigned based on the \code{community} parameter's chosen algorithm.
}
\examples{
# Example usage:
# Assuming 'x' is a tna object and 'community_assignment' holds the respective community labels
plot_communities(x, community_assignment, "walktrap")

}
