% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_model}
\alias{plot_model}
\title{Plot a Transition Network Model from a Matrix of Edge Weights}
\usage{
plot_model(x, labels, colors, cut, ...)
}
\arguments{
\item{x}{A square \code{matrix} of edge weights.}

\item{labels}{Optional \code{character} vector of node labels.
See \link[cograph:splot]{cograph::splot}.}

\item{colors}{An optional \code{character} vector of node colors to use.}

\item{cut}{A \code{numeric} value for the edge emphasis threshold.
See \link[cograph:splot]{cograph::splot}.}

\item{...}{Additional arguments passed to \code{\link[cograph:splot]{cograph::splot()}} or
\code{\link[cograph:plot_htna]{cograph::plot_htna()}}.}
}
\value{
See \code{\link[=plot.tna]{plot.tna()}}.
}
\description{
Plot a Transition Network Model from a Matrix of Edge Weights
}
\examples{
m <- matrix(rexp(25), 5, 5)
plot_model(m)

}
\keyword{internal}
