% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_model}
\alias{plot_model}
\title{Plot a Transition Network Model from a Matrix of Edge Weights}
\usage{
plot_model(
  x,
  labels,
  colors,
  cut,
  edge.labels = TRUE,
  edge.label.position = 0.65,
  layout = "circle",
  layout_args = list(),
  mar = rep(5, 4),
  theme = "colorblind",
  ...
)
}
\arguments{
\item{x}{A square \code{matrix} of edge weights.}

\item{...}{Additional arguments passed to \code{\link[Sonnet:plot_tna]{Sonnet::plot_tna()}} or
\code{\link[Sonnet:plot_htna]{Sonnet::plot_htna()}}.}
}
\value{
See \code{\link[=plot.tna]{plot.tna()}}.
}
\description{
Plot a Transition Network Model from a Matrix of Edge Weights
}
\examples{
m <- matrix(rexp(25), 5, 5)
plot_model(m)

}
\keyword{internal}
