% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.group_tna_centralities}
\alias{print.group_tna_centralities}
\title{Print Centrality Measures}
\usage{
\method{print}{group_tna_centralities}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna_centralities} object.}

\item{...}{Ignored.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Centrality Measures
}
\examples{
model <- group_model(engagement_mmm)
cm <- centralities(model)
print(cm)

}
\seealso{
Centrality measure functions
\code{\link{betweenness_network}()},
\code{\link{centralities}()},
\code{\link{plot.group_tna_centralities}()},
\code{\link{plot.tna_centralities}()},
\code{\link{print.tna_centralities}()}
}
\concept{centralities}
