% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.group_tna_cliques}
\alias{print.group_tna_cliques}
\title{Print Found Cliques}
\usage{
\method{print}{group_tna_cliques}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna_cliques} object.}

\item{...}{Arguments passed to \code{\link[=print.tna_cliques]{print.tna_cliques()}}.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Found Cliques
}
\examples{
model <- group_model(engagement_mmm)
cliq <- cliques(model, size = 2)
print(cliq)

}
\seealso{
Clique-related functions
\code{\link{cliques}()},
\code{\link{plot.group_tna_cliques}()},
\code{\link{plot.tna_cliques}()},
\code{\link{print.tna_cliques}()}
}
\concept{cliques}
