% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.group_tna_communities}
\alias{print.group_tna_communities}
\title{Print Detected Communities}
\usage{
\method{print}{group_tna_communities}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna_communities} object.}

\item{...}{Arguments passed to \code{\link[=print.tna_communities]{print.tna_communities()}}.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Detected Communities
}
\examples{
model <- group_model(engagement_mmm)
comm <- communities(model)
print(comm)

}
\seealso{
Community detection functions
\code{\link{communities}()},
\code{\link{plot.group_tna_communities}()},
\code{\link{plot.tna_communities}()},
\code{\link{print.tna_communities}()}
}
\concept{communities}
