% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.tna}
\alias{print.summary.tna}
\title{Print a TNA Summary}
\usage{
\method{print}{summary.tna}(x, ...)
}
\arguments{
\item{x}{A \code{summary.tna} object.}

\item{...}{Ignored.}
}
\value{
A \code{summary.tna} object (invisibly) containing the TNA model network
metrics and values.
}
\description{
Print a TNA Summary
}
\examples{
model <- tna(group_regulation)
print(summary(model))

}
\seealso{
Basic functions
\code{\link{build_model}()},
\code{\link{hist.group_tna}()},
\code{\link{hist.tna}()},
\code{\link{import_data}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.tna}()},
\code{\link{plot_frequencies}()},
\code{\link{plot_frequencies.group_tna}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_mosaic.group_tna}()},
\code{\link{plot_mosaic.tna_data}()},
\code{\link{prepare_data}()},
\code{\link{print.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.tna}()},
\code{\link{print.tna_data}()},
\code{\link{simulate.tna}()},
\code{\link{summary.group_tna}()},
\code{\link{summary.tna}()},
\code{\link{tna-package}}
}
\concept{basic}
