% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.tna_bootstrap}
\alias{print.summary.tna_bootstrap}
\title{Print Bootstrap Summary}
\usage{
\method{print}{summary.tna_bootstrap}(x, ...)
}
\arguments{
\item{x}{A \code{summary.tna_bootstrap} object.}

\item{...}{Arguments passed to the generic \code{print} method.}
}
\value{
A \code{summary.tna_bootstrap} object (invisibly) containing the weight,
p-value and confidence interval of each edge.
}
\description{
Print Bootstrap Summary
}
\examples{
model <- tna(group_regulation)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 10)
print(summary(boot))

}
