% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna}
\alias{print.tna}
\title{Print a \code{tna} object}
\usage{
\method{print}{tna}(x, digits = 2, generic = FALSE, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{digits}{Number of decimal digits to print. Defaults to 2.}

\item{generic}{Use generic print. Defaults to \code{FALSE}}

\item{...}{Ignored.}
}
\description{
Print a \code{tna} object
}
