% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna}
\alias{print.tna}
\title{Print a \code{tna} object}
\usage{
\method{print}{tna}(x, ...)
}
\arguments{
\item{x}{A \code{centralities} object.}

\item{...}{Ignored.}
}
\description{
Print a \code{tna} object
}
