% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna}
\alias{print.tna}
\title{Print \code{tna} object}
\usage{
\method{print}{tna}(x, generic = FALSE, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{generic}{Use generic print. Defaults to \code{FALSE}}

\item{...}{Ignored.}
}
\description{
Print \code{tna} object
}
