% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna}
\alias{print.tna}
\title{Print a \code{tna} object}
\usage{
\method{print}{tna}(x, digits = getOption("digits"), generic = FALSE, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{digits}{An \code{integer} giving the number of
\emph{significant} digits to print.}

\item{generic}{A \code{logical} value. If \code{TRUE}, use generic print method
instead. Defaults to \code{FALSE}.}

\item{...}{Ignored.}
}
\value{
The \code{tna} object passed as argument \code{x} (invisibly).
}
\description{
Print a \code{tna} object
}
\examples{
model <- tna(engagement)
print(model)

}
