% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_centralities}
\alias{print.tna_centralities}
\title{Print Centrality Measures}
\usage{
\method{print}{tna_centralities}(x, ...)
}
\arguments{
\item{x}{A \code{centralities} object.}

\item{...}{Ignored.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Centrality Measures
}
\examples{
model <- tna(group_regulation)
cm <- centralities(model)
print(cm)

}
