% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_cliques}
\alias{print.tna_cliques}
\title{Print Found Cliques}
\usage{
\method{print}{tna_cliques}(x, n = 6, first = 1, digits = 3, ...)
}
\arguments{
\item{x}{A \code{tna_cliques} object.}

\item{n}{An \code{integer} defining the maximum number of cliques to show.
The defaults is 6.}

\item{first}{An \code{integer} giving the index of the first clique to show.
The default index is 1.}

\item{digits}{An \code{integer} giving the number of digits to show for the edge
weights in the cliques. The default is 3.}

\item{...}{Ignored.}
}
\description{
Print Found Cliques
}
