% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_communities}
\alias{print.tna_communities}
\title{Print Detected Communities}
\usage{
\method{print}{tna_communities}(x, ...)
}
\arguments{
\item{x}{A \code{tna_communities} object.}

\item{...}{Ignored.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Detected Communities
}
\examples{
model <- tna(engagement)
comm <- communities(model)
print(comm)

}
