% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_mmm}
\alias{print.tna_mmm}
\title{Print a Mixed Markov Model Fit}
\usage{
\method{print}{tna_mmm}(x, ...)
}
\arguments{
\item{x}{A \code{tna_mmm} object.}

\item{...}{(Ignored)}
}
\value{
\code{x} (invisibly)
}
\description{
Print a Mixed Markov Model Fit
}
\examples{
\dontrun{
model <- cluster_mmm(engagement, k = 3)
print(model)
}

}
