% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tna_permutation}
\alias{print.tna_permutation}
\title{Print Permutation Test Results}
\usage{
\method{print}{tna_permutation}(x, ...)
}
\arguments{
\item{x}{A \code{tna_permutation} object.}

\item{...}{Additional arguments passed to the \code{tibble} print method.}
}
\value{
\code{x} (invisibly).
}
\description{
Print Permutation Test Results
}
\examples{
model_x <- tna(group_regulation[1:200, ])
model_y <- tna(group_regulation[1001:1200, ])
# Small number of iterations for CRAN
perm <- permutation_test(model_x, model_y, iter = 20)
print(perm)

}
