% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{prune.group_tna}
\alias{prune.group_tna}
\title{Prune a grouped Transition Network Analysis Model}
\usage{
\method{prune}{group_tna}(x, ...)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{...}{Same as \code{prune.tna}}
}
\value{
A \code{group_tna} object
}
\description{
Prune a grouped Transition Network Analysis Model
}
\seealso{
Cluster-related functions
\code{\link{as.igraph.group_tna}()},
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.summary.group_tna}()},
\code{\link{pruning_details.group_tna}()},
\code{\link{reprune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
