% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{prune.group_tna}
\alias{prune.group_tna}
\title{Prune a \code{group_tna} network based on transition probabilities}
\usage{
\method{prune}{group_tna}(x, ...)
}
\arguments{
\item{x}{An object of class \code{group_tna}.}

\item{...}{Arguments passed to \code{\link[=prune.tna]{prune.tna()}}.}
}
\value{
A pruned \code{group_tna} object. Details on the pruning can be viewed
with \code{\link[=pruning_details]{pruning_details()}}. The original model can be
restored with \code{\link[=deprune]{deprune()}}.
}
\description{
Prunes a set of networks represented by a \code{group_tna} object by removing
edges based on a specified threshold, lowest percent of non-zero edge
weights, or the disparity filter algorithm (Serrano et al., 2009).
It ensures the networks remain weakly connected.
}
\examples{
group <- c(rep("High", 100), rep("Low", 100))
model <- group_model(engagement, group = group)
pruned_threshold <- prune(model, method = "threshold", threshold = 0.1)
pruned_percentile <- prune(model,method = "lowest", lowest = 0.05)
pruned_disparity <- prune(model, method = "disparity", level = 0.5)

}
\seealso{
Cluster-related functions
\code{\link{bootstrap}()},
\code{\link{centralities}()},
\code{\link{cliques}()},
\code{\link{communities}()},
\code{\link{deprune}()},
\code{\link{estimate_cs}()},
\code{\link{group_model}()},
\code{\link{hist.group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{plot.group_tna_centralities}()},
\code{\link{plot.group_tna_cliques}()},
\code{\link{plot.group_tna_communities}()},
\code{\link{plot.group_tna_stability}()},
\code{\link{print.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_cliques}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{rename_groups}()},
\code{\link{reprune}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
