% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{prune}
\alias{prune}
\title{Prune a \code{tna} network based on transition probabilities}
\usage{
prune(x, threshold = NULL, percentile = NULL, lowest_percent = NULL)
}
\arguments{
\item{x}{An object of class \code{tna}}

\item{threshold}{A numeric value specifying the edge weight threshold. Edges with weights below or equal to this threshold will be considered for removal.}

\item{percentile}{A numeric value specifying the percentile of non-zero edges. Edges with weights below or equal to this percentile will be considered for removal.}

\item{lowest_percent}{A numeric value specifying the lowest percentage of non-zero edges. This percentage of edges with the lowest weights will be considered for removal.}
}
\value{
A list containing:
\describe{
\item{\code{pruned}}{The pruned \code{tna} object with updated transition matrix.}
\item{\code{removed_edges}}{A data frame of edges that were removed.}
\item{\code{num_removed_edges}}{The number of edges removed.}
\item{\code{threshold_used}}{The threshold value used for pruning.}
\item{\code{method_used}}{The method used to determine the threshold.}
}
}
\description{
This function prunes a network represented by a \code{tna} object by removing edges based on a specified threshold, percentile, or lowest percent of non-zero edge weights. It ensures the network remains weakly connected.
}
\details{
The function prunes the edges of the network(s) based on the specified threshold, percentile, or lowest percent of non-zero edge weights. It ensures that removing an edge does not disconnect the network (i.e., the network remains weakly connected).
\describe{
\item{\code{threshold}}{If specified, transition probabilities that are less than or equal to this value will be considered for removal.}
\item{\code{percentile}}{If specified, transition probabilities below this percentile of non-zero edge weights will be considered for removal.}
\item{\code{lowest_percent}}{If specified, the lowest percent of non-zero transition probabilities will be considered for removal.}
\item{\code{Default threshold}}{If no threshold, percentile, or lowest percent is specified, a default threshold of 0.05 is used.}
}
The function uses a depth-first search to ensure the network remains weakly connected after pruning.
}
\examples{
\dontrun{
  tna_model <- build_tna(engagement)
  pruned_result <- prune(tna_model, threshold = 0.1)
  pruned_result <- prune(tna_model, percentile = 10)
  pruned_result <- prune(tna_model, lowest_percent = 5)
}
}
\seealso{
Evaluation and validation functions
\code{\link{bootstrap_tna}()},
\code{\link{compare_tna_networks}()}
}
\author{
Mohammed Saqr (\email{mohammed.saqr@uef.fi})
}
\concept{evaluation}
