% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{prune}
\alias{prune}
\title{Prune a \code{tna} network based on transition probabilities}
\usage{
prune(
  x,
  method = c("threshold", "lowest", "disparity"),
  threshold = 0.1,
  lowest = 0.05,
  alpha = 0.5
)
}
\arguments{
\item{x}{An object of class \code{tna}}

\item{method}{A \code{character} string describing the pruning method.
The available options are \code{"threshold"}, \code{"lowest"}, \code{"percentile"}, and
\code{"disparity"}, corresponding to the methods listed in Details. The default
is \code{"threshold"}.}

\item{threshold}{A numeric value specifying the edge weight threshold.
Edges with weights below or equal to this threshold will be considered for
removal.}

\item{lowest}{A \code{numeric} value specifying the lowest percentage
of non-zero edges. This percentage of edges with the lowest weights will be
considered for removal. The default is \code{0.05}.}

\item{alpha}{A \code{numeric} value representing the significance level for the
disparity filter. Defaults to \code{0.5}.}
}
\value{
A list containing:
\itemize{
\item \code{pruned} The pruned \code{tna} object with updated transition matrix.
\item \code{removed_edges} A data frame of edges that were removed.
\item \code{num_removed_edges} The number of edges removed.
}
}
\description{
This function prunes a network represented by a \code{tna} object by removing
edges based on a specified threshold, lowest percent of non-zero edge
weights, or the disparity algorithm TODO cite. It ensures the
network remains weakly connected.
}
\examples{
\dontrun{
  tna_model <- build(engagement)
  pruned_threshold <- prune(tna_model, method = "threshold", threshold = 0.1)
  pruned_percentile <- prune(tna_model,method = "lowest", lowest = 0.05)
  pruned_disparity <- prune(tna_model, "disparity", alpha = 0.5)
}
}
\seealso{
Evaluation and validation functions
\code{\link{bootstrap}()},
\code{\link{compare_tna_networks}()}
}
\concept{evaluation}
