% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{prune}
\alias{prune}
\title{Prune a \code{tna} network based on transition probabilities}
\usage{
prune(x, ...)
}
\arguments{
\item{x}{An object of class \code{tna} or \code{group_tna}}
}
\description{
Prunes a network represented by a \code{tna} object by removing
edges based on a specified threshold, lowest percent of non-zero edge
weights, or the disparity filter algorithm (Serrano et al., 2009).
It ensures the network remains weakly connected.
}
\seealso{
Evaluation and validation functions
\code{\link{bootstrap}()},
\code{\link{permutation_test}()},
\code{\link{prune.tna}()}
}
\concept{evaluation}
