% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_details}
\alias{pruning_details}
\alias{pruning_details.tna}
\title{Print Detailed Information on the Pruning Results}
\usage{
pruning_details(x, ...)

\method{pruning_details}{tna}(x, removed_edges = TRUE, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{...}{Ignored.}

\item{removed_edges}{Should a \code{data.frame} of removed edges be printed?
The default is \code{FALSE}.}
}
\description{
Print Detailed Information on the Pruning Results
}
