% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R, R/pruning.R
\name{pruning_details.group_model}
\alias{pruning_details.group_model}
\alias{pruning_details}
\alias{pruning_details.tna}
\title{Print Detailed Information on the Pruning Results}
\usage{
\method{pruning_details}{group_model}(x, ...)

pruning_details(x, ...)

\method{pruning_details}{tna}(x, removed_edges = TRUE, ...)
}
\arguments{
\item{x}{A \code{tna} or \code{group_tna} object.}

\item{...}{Ignored.}

\item{removed_edges}{Should a \code{data.frame} of removed edges be printed?
The default is \code{FALSE}.}
}
\description{
Print Detailed Information on the Pruning Results

Print Detailed Information on the Pruning Results
}
\seealso{
Evaluation and validation functions
\code{\link{bootstrap}()},
\code{\link{permutation_test}()},
\code{\link{prune}()},
\code{\link{prune.tna}()}
}
\concept{evaluation}
