% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{rename_groups}
\alias{rename_groups}
\title{Rename clusters}
\usage{
rename_groups(x, new_names)
}
\arguments{
\item{x}{A \code{group_model} object}

\item{new_names}{A vector containing one name per cluster}
}
\value{
A renamed \code{group_model} object
}
\description{
Rename clusters
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_model}()},
\code{\link{deprune.group_model}()},
\code{\link{estimate_cs}()},
\code{\link{group_model}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_model}()},
\code{\link{print.group_model_bootstrap}()},
\code{\link{print.group_model_centralities}()},
\code{\link{print.group_model_communities}()},
\code{\link{print.summary.group_model}()},
\code{\link{print.summary.group_model_bootstrap}()},
\code{\link{prune.group_model}()},
\code{\link{reprune.group_model}()},
\code{\link{summary.group_model}()}
}
\concept{clusters}
