% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{rename_groups}
\alias{rename_groups}
\title{Rename Clusters}
\usage{
rename_groups(x, new_names)
}
\arguments{
\item{x}{A \code{group_tna} object.}

\item{new_names}{A \code{character} vector containing one name per cluster.}
}
\value{
A renamed \code{group_tna} object.
}
\description{
Rename Clusters
}
\examples{
model <- group_model(engagement_mmm)
model_renamed <- rename_groups(model, c("A", "B", "C"))

}
\seealso{
Cluster-related functions
\code{\link{communities}()},
\code{\link{group_model}()},
\code{\link{mmm_stats}()}
}
\concept{clusters}
