% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R, R/pruning.R
\name{reprune.group_tna}
\alias{reprune.group_tna}
\alias{reprune}
\title{Restore Previous Pruning of a Transition Network Analysis Model}
\usage{
\method{reprune}{group_tna}(x, ...)

reprune(x, ...)
}
\arguments{
\item{x}{A \code{tna} object.}

\item{...}{Ignored.}
}
\value{
A \code{tna} object that has not been pruned. The previous pruning
result can be reactivated with \code{\link[=reprune]{reprune()}}.
}
\description{
Restore Previous Pruning of a Transition Network Analysis Model
}
\examples{
model <- tna(engagement)
pruned_model <- prune(model, method = "threshold", threshold = 0.1)
depruned_model <- deprune(pruned_model) # restore original model
repruned_model <- reprune(depruned_model) # reapply the previous pruning
}
\seealso{
Cluster-related functions
\code{\link{centralities}()},
\code{\link{communities.group_tna}()},
\code{\link{deprune.group_tna}()},
\code{\link{estimate_cs}()},
\code{\link{group_tna}()},
\code{\link{mmm_stats}()},
\code{\link{plot.group_tna}()},
\code{\link{print.group_tna_bootstrap}()},
\code{\link{print.group_tna_centralities}()},
\code{\link{print.group_tna_communities}()},
\code{\link{print.summary.group_tna}()},
\code{\link{print.summary.group_tna_bootstrap}()},
\code{\link{prune.group_tna}()},
\code{\link{summary.group_tna}()}
}
\concept{clusters}
