% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{sequence_indices}
\alias{sequence_indices}
\alias{sequence_indices.tna}
\alias{sequence_indices.default}
\title{Compute Sequence Indices}
\usage{
sequence_indices(x, ...)

\method{sequence_indices}{tna}(x, favorable, omega = 1, ...)

\method{sequence_indices}{default}(x, cols, favorable, omega = 1, ...)
}
\arguments{
\item{x}{A \code{data.frame}, \code{matrix} or a \code{tna} object created from sequence
data.}

\item{...}{Not used.}

\item{favorable}{A \code{character} vector of state names that should be
considered as favorable states.}

\item{omega}{A \code{numeric} value for the omega parameter used to compute
the integrative potential.}

\item{cols}{An \code{expression} giving a tidy selection of columns that should
be considered as sequence data. By default, all columns are used.
Ignored if \code{x} is a \code{tna} object.}
}
\value{
A \code{data.frame} containing the index values.
}
\description{
Compute Sequence Indices
}
\examples{
indices <- sequence_indices(engagement)
print(indices)

}
