% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{sequence_indices}
\alias{sequence_indices}
\alias{sequence_indices.default}
\title{Compute Sequence Indices}
\usage{
sequence_indices(x, ...)

\method{sequence_indices}{default}(x, cols, favorable, min_length = 1L, omega = 1, ...)
}
\arguments{
\item{x}{A \code{data.frame} or a \code{matrix}.}
}
\value{
A \code{data.frame} containing the index values.
}
\description{
Compute Sequence Indices
}
