% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tna_bootstrap}
\alias{summary.tna_bootstrap}
\title{Summarize Bootstrap Results}
\usage{
\method{summary}{tna_bootstrap}(object, ...)
}
\arguments{
\item{object}{A \code{tna_bootstrap} object from \code{\link[=bootstrap]{bootstrap()}}.}

\item{...}{Ignored.}
}
\value{
A \code{summary.tna_bootstrap} object containing the weight,
p-value and confidence interval of each edge.
}
\description{
Summarize Bootstrap Results
}
\examples{
model <- tna(engagement)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 50)
summary(boot)

}
