% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tna_data}
\alias{summary.tna_data}
\title{Summarize TNA data converted from \code{prepare_data} or \code{prepare_ts}}
\usage{
\method{summary}{tna_data}(object, ...)
}
\arguments{
\item{object}{A \code{tna_data} object.}

\item{...}{Ignored.}
}
\value{
A \code{summary.tna_data} object containing the ....
}
\description{
Summarize TNA data converted from \code{prepare_data} or \code{prepare_ts}
}
\examples{
data <- prepare_data(group_regulation_long, action = "Action",
actor = "Actor", time = "Time")
summary(data)

}
\concept{data}
