% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{package}
\name{tna-package}
\alias{tna-package}
\title{The \code{tna} package.}
\description{
Provides tools for performing transition network analysis (TNA),
including functions for building TNA models, plotting transition networks,
and calculating centrality measures. The package relies on the \code{qgraph}
and \code{igraph} for network plotting and centrality measure calculations.
}
\references{
Saqr M., López-Pernas S., Törmänen T., Kaliisa R., Misiejuk K., Tikka S. (2025).
Transition Network Analysis: A Novel Framework for Modeling, Visualizing,
and Identifying the Temporal Patterns of Learners and Learning Processes.
In \emph{Proceedings of the 15th International Learning Analytics and Knowledge Conference (LAK '25)}, 351-361.

Banerjee A., Chandrasekhar A., Duflo E., Jackson M. (2014).
Gossip: Identifying Central Individuals in a Social Network.
Working Paper.

Kivimaki, I., Lebichot, B., Saramaki, J., Saerens, M. (2016).
Two betweenness centrality measures based on Randomized Shortest Paths.
\emph{Scientific Reports}, 6, 19668.

Serrano, M. A., Boguna, M., Vespignani, A. (2009).
Extracting the multiscale backbone of complex weighted networks.
\emph{Proceedings of the National Academy of Sciences}, 106, 6483-6488.

Zhang, B., Horvath, S. (2005).
A general framework for weighted gene co-expression network analysis.
\emph{Statistical Applications in Genetics and Molecular Biology}, 4(1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sonsoleslp/tna/}
  \item \url{http://sonsoles.me/tna/}
  \item Report bugs at \url{https://github.com/sonsoleslp/tna/issues/}
}

}
\author{
Sonsoles López-Pernas, Santtu Tikka, Mohammed Saqr
}
