% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna-package.R
\docType{package}
\name{tna-package}
\alias{tna}
\alias{tna-package}
\title{The \code{tna} package.}
\description{
Provides tools for performing transition network analysis (TNA),
including functions for building TNA models, plotting transition networks,
and calculating centrality measures. The package relies on the \code{qgraph}
and \code{igraph} for network plotting and centrality measure calculations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whitewingeddove/tna/}
  \item Report bugs at \url{https://github.com/whitewingeddove/tna/issues/}
}

}
\author{
\strong{Maintainer}: Anonymous \email{anon@anon.com}

Authors:
\itemize{
  \item Anon
  \item Anon
}

}
