% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tna.R
\name{tna}
\alias{tna}
\alias{tna.default}
\alias{tna.matrix}
\alias{tna.stslist}
\alias{tna.data.frame}
\alias{tna.mhmm}
\title{Build a Transition Network Analysis Model}
\usage{
tna(x, ...)

\method{tna}{default}(x, type = "prop", inits, ...)

\method{tna}{matrix}(x, type = "prop", inits, ...)

\method{tna}{stslist}(x, type = "prop", ...)

\method{tna}{data.frame}(x, type = "prop", ...)

\method{tna}{mhmm}(x, type = "prop", ...)
}
\arguments{
\item{x}{A \code{stslist} object describing a sequence of events or states to
be used for building
the Markov model or a \code{matrix} of transition probabilities with column
names describing the states. If \code{x} is a matrix, it is assumed that the
element on row \code{i} and column \code{j} is the transition probability (or weight)
from state \code{i} to state \code{j}. It also accepts a \code{data.frame} object in wide format
.  (each column is a timepoint with no extra columns)}

\item{...}{Ignored.}

\item{type}{A \code{character} string describing the weight matrix type.
Currently supports \code{"prop"} for proportions (probabilities) and
\code{"freq"} for frequencies.}

\item{inits}{An optional \code{numeric} vector of initial state probabilities
for each state. Can be provided only if \code{x} is a \code{matrix}. The vector will
be scaled to unity.}
}
\value{
An object of class \code{tna} which is a \code{list} containing the
following elements:
\itemize{
\item \code{weights}\cr A \code{list} of adjacency matrices of the model
(weight matrices) for each cluster.
\item \code{inits}\cr A \code{list} of initial weight vectors for each
cluster. For matrices, this element will be \code{NULL} if \code{inits} is not
directly provided
\item \code{labels}\cr A \code{character} vector of the state labels, or \code{NULL} if there
are no labels.
\item \code{seq}\cr The original sequence data converted to an internal format
used by the package when \code{x} is a \code{stslist} or a \code{data.frame} object.
}
}
\description{
This function constructs a transition network analysis (TNA) model from a
given sequence. It takes a sequence of events or states and builds a Markov
model. It extracts the transition probabilities
and initial probabilities from the model and stores them in a \code{list} along
with the state labels. Additionally, it creates a transition matrix with
zero diagonal entries (without loops). Also accepts matrices of transition
probabilities and initial state probabilities directly.
}
\examples{
tna_model <- tna(engagement)
print(tna_model)

}
\seealso{
Core functions
\code{\link{centralities}()},
\code{\link{plot.tna}()},
\code{\link{plot.tna_centralities}()},
\code{\link{plot_compare}()}
}
\concept{core}
