% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R, R/vcov.R
\name{coef.tna_mmm}
\alias{coef.tna_mmm}
\alias{vcov.tna_mmm}
\alias{vcov.summary.tna_mmm}
\title{Extract Regression Coefficients of a Mixture Markov Model}
\usage{
\method{coef}{tna_mmm}(object, ...)

\method{vcov}{tna_mmm}(object, ...)

\method{vcov}{summary.tna_mmm}(object, ...)
}
\arguments{
\item{object}{A \code{tna_mmm} or a \code{summary.tna_mmm} object.}

\item{...}{Not used.}
}
\value{
A \code{matrix} containing the regression coefficient estimates

A \code{matrix} containing the variance-covariance matrix.
}
\description{
Extract Regression Coefficients of a Mixture Markov Model

Variance-Covariance Matrix of the Covariate Coefficients of a Mixture Markov Model
}
\examples{
# TODO
}
